Split src/rt/cover.c into multiple files
