+++ Makefile.am
@@ -44,6 +44,7 @@ pkglib_LIBRARIES =
 
 include thirdparty/Makemodule.am
 include src/Makemodule.am
+include src/cov/Makemodule.am
 include src/psl/Makemodule.am
 include src/rt/Makemodule.am
 include src/jit/Makemodule.am
