+++ src/elab.c
@@ -18,6 +18,7 @@
 #include "util.h"
 #include "array.h"
 #include "common.h"
+#include "cov/cov-api.h"
 #include "diag.h"
 #include "driver.h"
 #include "eval.h"
@@ -30,7 +31,6 @@
 #include "option.h"
 #include "phase.h"
 #include "psl/psl-phase.h"
-#include "rt/cover.h"
 #include "type.h"
 #include "vlog/vlog-node.h"
 #include "vlog/vlog-phase.h"
