+++ src/rt/Makemodule.am
@@ -4,7 +4,6 @@ lib_libnvc_a_SOURCES += \
 	src/rt/wave.c \
 	src/rt/wave.h \
 	src/rt/rt.h \
-	src/rt/cover.h \
 	src/rt/heap.h \
 	src/rt/mspace.h \
 	src/rt/mspace.c \
