+++ src/rt/cover.c
@@ -18,7 +18,8 @@
 #include "util.h"
 #include "array.h"
 #include "common.h"
-#include "cover.h"
+#include "cov/cov-api.h"
+#include "cov/cov-data.h"
 #include "hash.h"
 #include "lib.h"
 #include "option.h"
@@ -31,184 +32,6 @@
 #include <stdlib.h>
 #include <string.h>
 #include <limits.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <unistd.h>
-#include <time.h>
-#include <ctype.h>
-#include <libgen.h>
-#include <time.h>
-
-//#define COVER_DEBUG_EMIT
-//#define COVER_DEBUG_DUMP
-//#define COVER_DEBUG_SCOPE
-//#define COVER_DEBUG_MERGE
-//#define COVER_DEBUG_CALLBACK
-//#define COVER_DEBUG_EXCLUDE
-
-#define MARGIN_LEFT "20%%"
-#define SIDEBAR_WIDTH "15%%"
-
-typedef struct {
-   int start;
-   int end;
-} line_range_t;
-
-typedef A(cover_item_t) cov_item_array_t;
-typedef A(line_range_t) range_array_t;
-typedef A(char*) char_array_t;
-
-typedef struct _cover_report_ctx    cover_report_ctx_t;
-typedef struct _cover_file          cover_file_t;
-typedef struct _cover_scope         cover_scope_t;
-typedef struct _cover_exclude_ctx   cover_exclude_ctx_t;
-typedef struct _cover_rpt_buf       cover_rpt_buf_t;
-typedef struct _cover_spec          cover_spec_t;
-
-typedef enum {
-   CSCOPE_UNKNOWN,
-   CSCOPE_INSTANCE,
-} scope_type_t;
-
-typedef struct _cover_scope {
-   scope_type_t      type;
-   ident_t           name;
-   ident_t           hier;
-   loc_t             loc;
-   int               branch_label;
-   int               stmt_label;
-   int               expression_label;
-   cover_scope_t    *parent;
-   ptr_list_t        children;
-   cov_item_array_t  items;
-   range_array_t     ignore_lines;
-   ident_t           block_name;
-   int               sig_pos;
-   bool              emit;
-} cover_scope_t;
-
-struct _cover_rpt_buf {
-   text_buf_t      *tb;
-   cover_rpt_buf_t *prev;
-};
-
-struct _cover_spec {
-   char_array_t hier_include;
-   char_array_t hier_exclude;
-   char_array_t block_include;
-   char_array_t block_exclude;
-   char_array_t fsm_type_include;
-   char_array_t fsm_type_exclude;
-};
-
-struct _cover_data {
-   int               next_tag;
-   cover_mask_t      mask;
-   int               array_limit;
-   int               array_depth;
-   int               report_item_limit;
-   cover_rpt_buf_t  *rpt_buf;
-   cover_spec_t     *spec;
-   cover_scope_t    *top_scope;
-   cover_scope_t    *root_scope;
-};
-
-typedef struct {
-   unsigned    total_stmts;
-   unsigned    hit_stmts;
-   unsigned    total_branches;
-   unsigned    hit_branches;
-   unsigned    total_toggles;
-   unsigned    hit_toggles;
-   unsigned    total_expressions;
-   unsigned    hit_expressions;
-   unsigned    total_states;
-   unsigned    hit_states;
-} cover_stats_t;
-
-typedef struct {
-   char       *text;
-   size_t      len;
-} cover_line_t;
-
-typedef struct {
-   cover_line_t *line;
-   cover_item_t *item;
-   int flags;
-} cover_pair_t;
-
-typedef struct {
-   cover_pair_t *hits;
-   cover_pair_t *miss;
-   cover_pair_t *excl;
-   int          n_hits;
-   int          n_miss;
-   int          n_excl;
-   int          alloc_hits;
-   int          alloc_miss;
-   int          alloc_excl;
-} cover_chain_t;
-
-struct _cover_file {
-   const char   *name;
-   cover_line_t *lines;
-   unsigned      n_lines;
-   unsigned      alloc_lines;
-   bool          valid;
-   cover_file_t *next;
-};
-
-struct _cover_report_ctx {
-   cover_data_t         *data;
-   cover_stats_t        flat_stats;
-   cover_stats_t        nested_stats;
-   cover_report_ctx_t   *parent;
-   cover_chain_t        ch_stmt;
-   cover_chain_t        ch_branch;
-   cover_chain_t        ch_toggle;
-   cover_chain_t        ch_expression;
-   cover_chain_t        ch_state;
-   int                  lvl;
-};
-
-#define INIT_CHAIN(ctx, name)                                           \
-   ctx->name.hits = xcalloc_array(1024, sizeof(cover_pair_t));          \
-   ctx->name.miss = xcalloc_array(1024, sizeof(cover_pair_t));          \
-   ctx->name.excl = xcalloc_array(1024, sizeof(cover_pair_t));          \
-   ctx->name.alloc_hits = 1024;                                         \
-   ctx->name.alloc_miss = 1024;                                         \
-   ctx->name.alloc_excl = 1024;                                         \
-
-static cover_file_t  *files;
-
-struct _cover_exclude_ctx {
-   FILE    *ef;
-   char    *line;
-   loc_t    loc;
-};
-
-static const char* item_kind_str[] = {
-   "statement",
-   "branch",
-   "toggle",
-   "expression",
-   "FSM state"
-};
-
-static const struct {
-   const char *name;
-   uint32_t   flag;
-} bin_map[] = {
-   { "BIN_TRUE",       COV_FLAG_TRUE},
-   { "BIN_FALSE",      COV_FLAG_FALSE},
-   { "BIN_CHOICE",     COV_FLAG_CHOICE},
-   { "BIN_0_0",        COV_FLAG_00},
-   { "BIN_0_1",        COV_FLAG_01},
-   { "BIN_1_0",        COV_FLAG_10},
-   { "BIN_1_1",        COV_FLAG_11},
-   { "BIN_0_TO_1",     COV_FLAG_TOGGLE_TO_1},
-   { "BIN_1_TO_0",     COV_FLAG_TOGGLE_TO_0},
-};
 
 enum std_ulogic {
    _U  = 0x0,
@@ -222,860 +45,7 @@ enum std_ulogic {
    _DC = 0x8
 };
 
-typedef enum {
-   PAIR_UNCOVERED    = 0,
-   PAIR_EXCLUDED     = 1,
-   PAIR_COVERED      = 2,
-   PAIR_LAST         = 3
-} cov_pair_kind_t;
-
-typedef enum {
-   CTRL_PUSH_SCOPE,
-   CTRL_POP_SCOPE,
-   CTRL_END_OF_FILE,
-} cov_control_t;
-
-static void cover_report_children(cover_report_ctx_t *ctx,
-                                  cover_scope_t *s, const char *dir,
-                                  FILE *summf, int *skipped);
-
-bool cover_is_stmt(tree_t t)
-{
-   switch (tree_kind(t)) {
-   case T_IF:
-   case T_WHILE:
-   case T_NEXT:
-   case T_EXIT:
-   case T_SIGNAL_ASSIGN:
-   case T_ASSERT:
-   case T_VAR_ASSIGN:
-   case T_RETURN:
-   case T_FOR:
-   case T_PCALL:
-   case T_FCALL:
-   case T_CASE:
-      return true;
-
-   // Static waits are introduced during simp pass. These are hidden
-   // for user, no need to cover them.
-   case T_WAIT:
-      if (tree_flags(t) & TREE_F_STATIC_WAIT)
-         return false;
-      return true;
-
-   default:
-      return false;
-   }
-}
-
-static bool cover_is_branch(tree_t branch)
-{
-   return tree_kind(branch) == T_ASSOC || tree_kind(branch) == T_COND_STMT;
-}
-
-static bool cover_is_toggle_first(tree_t decl)
-{
-   const tree_kind_t kind = tree_kind(decl);
-   return kind == T_SIGNAL_DECL || kind == T_PORT_DECL;
-}
-
-unsigned cover_get_std_log_expr_flags(tree_t decl)
-{
-   if (tree_kind(decl) != T_FUNC_DECL)
-      return 0;
-
-   tree_t container = tree_container(decl);
-   if (is_well_known(tree_ident(container)) != W_IEEE_1164)
-      return 0;
-
-   struct {
-      well_known_t op;
-      unsigned flags;
-   } std_log_ops[] = {
-      { W_IEEE_1164_AND    , COVER_FLAGS_AND_EXPR},
-      { W_IEEE_1164_NAND   , COVER_FLAGS_AND_EXPR},
-      { W_IEEE_1164_OR     , COVER_FLAGS_OR_EXPR},
-      { W_IEEE_1164_NOR    , COVER_FLAGS_OR_EXPR},
-      { W_IEEE_1164_XOR    , COVER_FLAGS_XOR_EXPR},
-      { W_IEEE_1164_XNOR   , COVER_FLAGS_XOR_EXPR}
-   };
-
-   unsigned flags = 0;
-   for (int i = 0; i < ARRAY_LEN(std_log_ops); i++)
-      if (ident_starts_with(tree_ident2(decl), well_known(std_log_ops[i].op)))
-         flags |= std_log_ops[i].flags;
-
-   return flags;
-}
-
-bool cover_skip_array_toggle(cover_data_t *data, int a_size)
-{
-   assert (data);
-
-   // Array is equal to or than configured limit
-   if (data->array_limit != 0 && a_size >= data->array_limit)
-      return true;
-
-   // Array is multi-dimensional or nested
-   if ((!cover_enabled(data, COVER_MASK_TOGGLE_INCLUDE_MEMS)) && data->array_depth > 0)
-      return true;
-
-   return false;
-}
-
-bool cover_skip_type_state(cover_data_t *data, type_t type)
-{
-   if (!type_is_enum(type))
-      return true;
-
-   // Ignore enums from built-in libraries
-   ident_t full_name = type_ident(type);
-   if (ident_starts_with(full_name, well_known(W_STD)) ||
-       ident_starts_with(full_name, well_known(W_IEEE)) ||
-       ident_starts_with(full_name, well_known(W_NVC)) ||
-       ident_starts_with(full_name, well_known(W_VITAL)))
-      return true;
-
-   ident_t name = ident_rfrom(full_name, '.');
-   cover_spec_t *spc = data->spec;
-
-   // Type should be included
-   if (spc) {
-      for (int i = 0; i < spc->fsm_type_include.count; i++)
-         if (ident_glob(name, AGET(spc->fsm_type_include, i), -1)) {
-#ifdef COVER_DEBUG_EMIT
-            printf("Cover emit: True, fsm-type (Type: %s, Pattern: %s)\n",
-                  istr(name), AGET(spc->fsm_type_include, i));
-#endif
-            return false;
-         }
-   }
-
-   // By default enums should not included
-   if (data->mask & COVER_MASK_FSM_NO_DEFAULT_ENUMS)
-      return true;
-
-   // Type should not be included
-   if (spc) {
-      for (int i = 0; i < spc->fsm_type_exclude.count; i++)
-         if (ident_glob(name, AGET(spc->fsm_type_exclude, i), -1)) {
-#ifdef COVER_DEBUG_EMIT
-            printf("Cover emit: False, fsm-type (Type: %s, Pattern: %s)\n",
-                   istr(name), AGET(spc->fsm_type_exclude, i));
-#endif
-            return true;
-         }
-     }
-
-   return false;
-}
-
-fbuf_t *cover_open_lib_file(tree_t top, fbuf_mode_t mode, bool check_null)
-{
-   char *dbname LOCAL = xasprintf("_%s.covdb", istr(tree_ident(top)));
-   fbuf_t *f = lib_fbuf_open(lib_work(), dbname, mode, FBUF_CS_NONE);
-
-   if (check_null && (f == NULL))
-      fatal_errno("failed to open coverage db file: %s", dbname);
-
-   return f;
-}
-
-cover_item_t *cover_add_item(tree_t t, const loc_t *loc, ident_t suffix,
-                             cover_data_t *data, cover_item_kind_t kind,
-                             uint32_t flags)
-{
-   assert(data != NULL);
-
-   if (!data->top_scope->emit)
-      return NULL;
-
-   cover_scope_t *ignore_scope = data->top_scope;
-   for (; ignore_scope->type != CSCOPE_INSTANCE && ignore_scope->parent;
-        ignore_scope = ignore_scope->parent)
-      ;
-
-   for (int i = 0; i < ignore_scope->ignore_lines.count; i++) {
-      line_range_t *lr = &(ignore_scope->ignore_lines.items[i]);
-      if (loc->first_line > lr->start && loc->first_line <= lr->end)
-         return NULL;
-   }
-
-   // Everything creates scope, so name of current item is already given
-   // by scope in hierarchy.
-   ident_t hier = data->top_scope->hier;
-   if (suffix)
-      hier = ident_prefix(hier, suffix, '\0');
-
-   // Expression items do not nest scope, expression name must be created
-   if (kind == COV_ITEM_EXPRESSION) {
-      char buf[16];
-      checked_sprintf(buf, sizeof(buf), "_E%d", data->top_scope->expression_label);
-      hier = ident_prefix(hier, ident_new(buf), '.');
-      data->top_scope->expression_label++;
-   }
-
-   // Query LHS/RHS operand locations of binary expressions
-   loc_t loc_lhs = LOC_INVALID, loc_rhs = LOC_INVALID;
-   if (flags & COVER_FLAGS_LHS_RHS_BINS) {
-      assert(tree_params(t) > 1);
-      loc_lhs = *tree_loc(tree_param(t, 0));
-      loc_rhs = *tree_loc(tree_param(t, 1));
-   }
-
-   ident_t func_name = NULL;
-   if (kind == COV_ITEM_EXPRESSION)
-      func_name = tree_ident(t);
-   else if (kind == COV_ITEM_STATE)
-      func_name = ident_rfrom(type_ident(tree_type(t)), '.');
-
-#ifdef COVER_DEBUG_EMIT
-   printf("Item: %s\n", istr(hier));
-   printf("    First line: %d\n", tree_loc(t)->first_line);
-   printf("    First column: %d\n", tree_loc(t)->first_column);
-   printf("    Line delta: %d\n", tree_loc(t)->line_delta);
-   printf("    Column delta: %d\n", tree_loc(t)->column_delta);
-   printf("\n\n");
-#endif
-
-   int num = 0;
-   if (kind == COV_ITEM_STATE) {
-      type_t enum_type = tree_type(t);
-      assert(type_is_enum(enum_type));
-      int64_t low, high;
-      folded_bounds(range_of(enum_type, 0), &low, &high);
-      num = high - low + 1;
-   }
-   else
-      num = data->top_scope->sig_pos;
-
-   cover_item_t new = {
-      .kind       = kind,
-      .tag        = data->next_tag++,
-      .data       = 0,
-      .flags      = flags,
-      .excl_msk   = 0,
-      .unrc_msk   = 0,
-      .loc        = *loc,
-      .loc_lhs    = loc_lhs,
-      .loc_rhs    = loc_rhs,
-      .hier       = hier,
-      .tree_kind  = tree_kind(t),
-      .func_name  = func_name,
-      .num        = num,
-   };
-
-   APUSH(data->top_scope->items, new);
-
-   return AREF(data->top_scope->items, data->top_scope->items.count - 1);
-}
-
-LCOV_EXCL_START
-static void cover_debug_dump(cover_scope_t *s, int indent)
-{
-   color_printf("%*s$!blue$%s", indent, "", istr(s->name));
-   switch (s->type) {
-   case CSCOPE_INSTANCE: color_printf(" : instance"); break;
-   default: break;
-   }
-   color_printf("$$\n");
-
-   for (int i = 0; i < s->items.count; i++) {
-      cover_item_t *item = &(s->items.items[i]);
-
-      char *path LOCAL = xstrdup(loc_file_str(&item->loc));
-      printf("%*s%d: %s %s %s:%d => %x\n", indent + 2, "", item->tag,
-             item_kind_str[item->kind], istr(item->hier),
-             basename(path), item->loc.first_line, item->data);
-   }
-
-   for (list_iter(cover_scope_t *, c, s->children))
-      cover_debug_dump(c, indent + 2);
-}
-LCOV_EXCL_STOP
-
-static void cover_merge_one_item(cover_item_t *item, int32_t data)
-{
-   switch (item->kind) {
-   case COV_ITEM_STMT:
-      item->data += data;
-      break;
-   case COV_ITEM_TOGGLE:
-   case COV_ITEM_BRANCH:
-   case COV_ITEM_EXPRESSION:
-   case COV_ITEM_STATE:
-      item->data |= data;
-      break;
-   default:
-      break;
-   }
-}
-
-static void cover_update_counts(cover_scope_t *s, const int32_t *counts)
-{
-   for (int i = 0; i < s->items.count; i++) {
-      cover_item_t *item = &(s->items.items[i]);
-
-      const int32_t data = counts ? counts[item->tag] : 0;
-      cover_merge_one_item(item, data);
-   }
-
-   for (list_iter(cover_scope_t *, it, s->children))
-      cover_update_counts(it, counts);
-}
-
-static void cover_write_scope(cover_scope_t *s, fbuf_t *f,
-                              ident_wr_ctx_t ident_ctx, loc_wr_ctx_t *loc_ctx)
-{
-   write_u8(CTRL_PUSH_SCOPE, f);
-
-   fbuf_put_uint(f, s->type);
-   ident_write(s->name, ident_ctx);
-   ident_write(s->hier, ident_ctx);
-   loc_write(&s->loc, loc_ctx);
-
-   if (s->type == CSCOPE_INSTANCE)
-      ident_write(s->block_name, ident_ctx);
-
-   fbuf_put_uint(f, s->items.count);
-   for (int i = 0; i < s->items.count; i++) {
-      cover_item_t *item = &(s->items.items[i]);
-
-      write_u8(item->kind, f);
-      write_u32(item->tag, f);
-      write_u32(item->data, f);
-      write_u32(item->flags, f);
-      write_u32(item->tree_kind, f);
-      // Do not dump "excl_msk" since it is only filled at
-      // report generation time
-      write_u32(item->unrc_msk, f);
-
-      loc_write(&(item->loc), loc_ctx);
-      if (item->flags & COVER_FLAGS_LHS_RHS_BINS) {
-         loc_write(&(item->loc_lhs), loc_ctx);
-         loc_write(&(item->loc_rhs), loc_ctx);
-      }
-
-      ident_write(item->hier, ident_ctx);
-      if (item->kind == COV_ITEM_EXPRESSION || item->kind == COV_ITEM_STATE)
-         ident_write(item->func_name, ident_ctx);
-
-      write_u32(item->num, f);
-   }
-
-   for (list_iter(cover_scope_t *, it, s->children))
-      cover_write_scope(it, f, ident_ctx, loc_ctx);
-
-   write_u8(CTRL_POP_SCOPE, f);
-}
-
-void cover_dump_items(cover_data_t *data, fbuf_t *f, cover_dump_t dt,
-                      const int32_t *counts)
-{
-   if (dt == COV_DUMP_RUNTIME)
-      cover_update_counts(data->root_scope, counts);
-
-   if (opt_get_int(OPT_COVER_VERBOSE))
-      cover_debug_dump(data->root_scope, 0);
-
-   write_u32(data->mask, f);
-   write_u32(data->array_limit, f);
-   write_u32(data->next_tag, f);
-
-   loc_wr_ctx_t *loc_wr = loc_write_begin(f);
-   ident_wr_ctx_t ident_ctx = ident_write_begin(f);
-
-   cover_write_scope(data->root_scope, f, ident_ctx, loc_wr);
-
-   write_u8(CTRL_END_OF_FILE, f);
-
-   loc_write_end(loc_wr);
-   ident_write_end(ident_ctx);
-}
-
-cover_data_t *cover_data_init(cover_mask_t mask, int array_limit)
-{
-   cover_data_t *data = xcalloc(sizeof(cover_data_t));
-   data->mask = mask;
-   data->array_limit = array_limit;
-
-   return data;
-}
-
-bool cover_enabled(cover_data_t *data, cover_mask_t mask)
-{
-   return data != NULL && (data->mask & mask);
-}
-
-static bool cover_should_emit_scope(cover_data_t *data, tree_t t)
-{
-   cover_scope_t *ts = data->top_scope;
-   cover_spec_t *spc = data->spec;
-
-   // Block (entity, package instance or block) name
-   if (ts->block_name) {
-      ident_t ename = ident_until(ts->block_name, '-');
-
-      for (int i = 0; i < spc->block_exclude.count; i++)
-
-         if (ident_glob(ename, AGET(spc->block_exclude, i), -1)) {
-#ifdef COVER_DEBUG_EMIT
-            printf("Cover emit: False, block (Block: %s, Pattern: %s)\n",
-                   istr(ts->block_name), AGET(spc->block_exclude, i));
-#endif
-            return false;
-         }
-
-      for (int i = 0; i < data->spec->block_include.count; i++)
-         if (ident_glob(ename, AGET(spc->block_include, i), -1)) {
-#ifdef COVER_DEBUG_EMIT
-            printf("Cover emit: True, block (Block: %s, Pattern: %s)\n",
-                   istr(ts->block_name), AGET(spc->block_include, i));
-#endif
-            return true;
-         }
-   }
-
-   // Hierarchy
-   for (int i = 0; i < spc->hier_exclude.count; i++)
-      if (ident_glob(ts->hier, AGET(spc->hier_exclude, i), -1)) {
-#ifdef COVER_DEBUG_EMIT
-         printf("Cover emit: False, hierarchy (Hierarchy: %s, Pattern: %s)\n",
-                istr(ts->hier), AGET(spc->hier_exclude, i));
-#endif
-         return false;
-      }
-
-   for (int i = 0; i < spc->hier_include.count; i++)
-      if (ident_glob(ts->hier, AGET(spc->hier_include, i), -1)) {
-#ifdef COVER_DEBUG_EMIT
-         printf("Cover emit: True, hierarchy (Hierarchy: %s, Pattern: %s)\n",
-                istr(ts->hier), AGET(spc->hier_include, i));
-#endif
-         return true;
-      }
-
-   return false;
-}
-
-void cover_push_scope(cover_data_t *data, tree_t t)
-{
-   if (data == NULL)
-      return;
-   else if (data->root_scope == NULL) {
-      cover_scope_t *s = xcalloc(sizeof(cover_scope_t));
-      s->name = s->hier = lib_name(lib_work());
-
-      data->top_scope = data->root_scope = s;
-   }
-
-   cover_scope_t *s = xcalloc(sizeof(cover_scope_t));
-   ident_t name = NULL;
-   char prefix[16] = {0};
-   int *cnt;
-   char c = 0;
-
-   if (cover_is_branch(t)) {
-      if (tree_kind(t) == T_ASSOC && tree_subkind(t) == A_OTHERS)
-         checked_sprintf(prefix, sizeof(prefix), "_B_OTHERS");
-      else {
-         assert(data->top_scope);
-         cnt = &data->top_scope->branch_label;
-         c = 'B';
-      }
-   }
-   // For toggle coverage, remember the position where its name in
-   // the hierarchy starts.
-   else if (cover_is_toggle_first(t)) {
-      assert(tree_has_ident(t));
-      name = tree_ident(t);
-      s->sig_pos = ident_len(data->top_scope->hier) + 1;
-   }
-   else if (tree_has_ident(t))
-      name = tree_ident(t);
-   // Consider everything else as statement
-   // Expressions do not get scope pushed, so if scope for e.g.
-   // T_FCALL is pushed it will be concurent function call -> Label as statement
-   else {
-      assert(data->top_scope);
-      cnt = &data->top_scope->stmt_label;
-      c = 'S';
-   }
-
-   if (c) {
-      checked_sprintf(prefix, sizeof(prefix), "_%c%u", c, *cnt);
-      (*cnt)++;
-   }
-   if (name == NULL)
-      name = ident_new(prefix);
-
-   s->name       = name;
-   s->parent     = data->top_scope;
-   s->block_name = s->parent->block_name;
-   s->loc        = *tree_loc(t);
-   s->hier       = ident_prefix(s->parent->hier, name, '.');
-
-   if (s->sig_pos == 0)
-      s->sig_pos = data->top_scope->sig_pos;
-
-   list_add(&data->top_scope->children, s);
-
-   data->top_scope = s;
-
-   if (tree_kind(t) == T_BLOCK) {
-      tree_t hier = tree_decl(t, 0);
-      assert(tree_kind(hier) == T_HIER);
-
-      tree_t unit = tree_ref(hier);
-      if (tree_kind(unit) == T_ARCH) {
-         s->block_name = ident_rfrom(tree_ident(unit), '.');
-         s->type = CSCOPE_INSTANCE;
-      }
-   }
-
-   s->emit = (data->spec == NULL) ? true : cover_should_emit_scope(data, t);
-
-#ifdef COVER_DEBUG_SCOPE
-   printf("Pushing cover scope: %s\n", istr(s->hier));
-   printf("Tree_kind: %s\n\n", tree_kind_str(tree_kind(t)));
-   printf("Coverage emit: %d\n\n", s->emit);
-#endif
-}
-
-void cover_pop_scope(cover_data_t *data)
-{
-   if (data == NULL)
-      return;
-
-   assert(data->top_scope != NULL);
-
-   ACLEAR(data->top_scope->ignore_lines);
-
-#ifdef COVER_DEBUG_SCOPE
-   printf("Popping cover scope: %s\n", istr(data->top_scope->hier));
-#endif
-
-   data->top_scope = data->top_scope->parent;
-
-}
-
-void cover_ignore_from_pragmas(cover_data_t *data, tree_t unit)
-{
-   assert(data->top_scope != NULL);
-
-   if (!is_design_unit(unit))
-      return;   // Generate block, etc.
-
-   range_array_t *excl = &(data->top_scope->ignore_lines);
-   bool state = true;
-   const int npragmas = tree_pragmas(unit);
-   for (int i = 0; i < npragmas; i++) {
-      tree_t p = tree_pragma(unit, i);
-      const pragma_kind_t kind = tree_subkind(p);
-      if (kind != PRAGMA_COVERAGE_OFF && kind != PRAGMA_COVERAGE_ON)
-         continue;
-
-      if (kind == PRAGMA_COVERAGE_OFF && state) {
-         line_range_t lr = { tree_loc(p)->first_line, INT_MAX };
-         APUSH(*excl, lr);
-         state = false;
-      }
-      else if (kind == PRAGMA_COVERAGE_ON && !state) {
-         assert(excl->count > 0);
-         assert(excl->items[excl->count - 1].end == INT_MAX);
-         excl->items[excl->count - 1].end = tree_loc(p)->first_line;
-         state = true;
-      }
-   }
-}
-
-void cover_inc_array_depth(cover_data_t *data)
-{
-   assert(data != NULL);
-   data->array_depth++;
-#ifdef COVER_DEBUG_SCOPE
-   printf("Adding dimension: %d\n", data->array_depth);
-#endif
-}
-
-void cover_dec_array_depth(cover_data_t *data)
-{
-   assert(data != NULL);
-   assert(data->array_depth > 0);
-   data->array_depth--;
-#ifdef COVER_DEBUG_SCOPE
-   printf("Subtracting dimension: %d\n", data->array_depth);
-#endif
-}
-
-static void cover_read_header(fbuf_t *f, cover_data_t *data)
-{
-   assert(data != NULL);
-
-   data->mask = read_u32(f);
-   data->array_limit = read_u32(f);
-   data->next_tag = read_u32(f);
-}
-
-static void cover_read_one_item(fbuf_t *f, loc_rd_ctx_t *loc_rd,
-                                ident_rd_ctx_t ident_ctx, cover_item_t *item)
-{
-   item->kind = read_u8(f);
-
-   item->tag = read_u32(f);
-   item->data = read_u32(f);
-   item->flags = read_u32(f);
-   item->tree_kind = read_u32(f);
-   item->unrc_msk = read_u32(f);
-   item->excl_msk = 0;
-
-   loc_read(&(item->loc), loc_rd);
-   if (item->flags & COVER_FLAGS_LHS_RHS_BINS) {
-      loc_read(&(item->loc_lhs), loc_rd);
-      loc_read(&(item->loc_rhs), loc_rd);
-   }
-
-   item->hier = ident_read(ident_ctx);
-   if (item->kind == COV_ITEM_EXPRESSION || item->kind == COV_ITEM_STATE)
-      item->func_name = ident_read(ident_ctx);
-
-   item->num = read_u32(f);
-}
-
-static cover_scope_t *cover_read_scope(fbuf_t *f, ident_rd_ctx_t ident_ctx,
-                                       loc_rd_ctx_t *loc_ctx)
-{
-   cover_scope_t *s = xcalloc(sizeof(cover_scope_t));
-   s->type = fbuf_get_uint(f);
-   s->name = ident_read(ident_ctx);
-   s->hier = ident_read(ident_ctx);
-
-   loc_read(&s->loc, loc_ctx);
-
-   if (s->type == CSCOPE_INSTANCE)
-      s->block_name = ident_read(ident_ctx);
-
-   const int nitems = fbuf_get_uint(f);
-   for (int i = 0; i < nitems; i++) {
-      cover_item_t new;
-      cover_read_one_item(f, loc_ctx, ident_ctx, &new);
-
-      APUSH(s->items, new);
-   }
-
-   for (;;) {
-      const uint8_t ctrl = read_u8(f);
-      switch (ctrl) {
-      case CTRL_PUSH_SCOPE:
-         {
-            cover_scope_t *child = cover_read_scope(f, ident_ctx, loc_ctx);
-            list_add(&s->children, child);
-         }
-         break;
-      case CTRL_POP_SCOPE:
-         return s;
-      default:
-         fatal_trace("invalid control word %x in cover db", ctrl);
-      }
-   }
-}
-
-cover_data_t *cover_read_items(fbuf_t *f, uint32_t pre_mask)
-{
-   cover_data_t *data = xcalloc(sizeof(cover_data_t));
-   cover_read_header(f, data);
-   data->mask |= pre_mask;
-
-   loc_rd_ctx_t *loc_rd = loc_read_begin(f);
-   ident_rd_ctx_t ident_ctx = ident_read_begin(f);
-
-   bool eof = false;
-   do {
-      const uint8_t ctrl = read_u8(f);
-      switch (ctrl) {
-      case CTRL_PUSH_SCOPE:
-         data->root_scope = cover_read_scope(f, ident_ctx, loc_rd);
-         break;
-      case CTRL_END_OF_FILE:
-         eof = true;
-         break;
-      default:
-         fatal_trace("invalid control word %x in cover db", ctrl);
-      }
-   } while (!eof);
-
-   ident_read_end(ident_ctx);
-   loc_read_end(loc_rd);
-
-   return data;
-}
-
-static void cover_merge_scope(cover_scope_t *old_s, cover_scope_t *new_s)
-{
-   // TODO: Could merging be done more efficiently?
-   bool found = false;
-   for (int i = 0; i < new_s->items.count; i++) {
-      cover_item_t *new = AREF(new_s->items, i);
-
-      for (int j = 0; j < old_s->items.count; j++) {
-         cover_item_t *old = AREF(old_s->items, j);
-
-         // Compare based on hierarchical path, each
-         // statement / branch / signal has unique hierarchical name
-         if (new->hier == old->hier) {
-            assert(new->kind == old->kind);
-#ifdef COVER_DEBUG_MERGE
-            printf("Merging coverage item: %s\n", istr(old->hier));
-#endif
-            cover_merge_one_item(old, new->data);
-
-            found = true;
-            break;
-         }
-
-         // TODO: Append the new item just before popping hierarchy item
-         //       with longest common prefix of new item. That will allow to
-         //       merge coverage of IPs from different configurations of
-         //       generics which form hierarchy differently!
-         if (!found)
-            warnf("dropping coverage item: %s", istr(new->hier));
-      }
-   }
-
-   for (list_iter(cover_scope_t *, new_c, new_s->children)) {
-      bool found = false;
-      for (list_iter(cover_scope_t *, old_c, old_s->children)) {
-         if (new_c->name == old_c->name) {
-            cover_merge_scope(old_c, new_c);
-            found = true;
-            break;
-         }
-      }
-
-      if (!found)
-         warnf("scope %s not found in original database", istr(new_c->name));
-   }
-}
-
-void cover_merge_items(fbuf_t *f, cover_data_t *data)
-{
-   assert (data != NULL);
-
-   cover_read_header(f, data);
-
-   loc_rd_ctx_t *loc_rd = loc_read_begin(f);
-   ident_rd_ctx_t ident_ctx = ident_read_begin(f);
-
-   bool eof = false;
-   do {
-      const uint8_t ctrl = read_u8(f);
-      switch (ctrl) {
-      case CTRL_PUSH_SCOPE:
-         {
-            cover_scope_t *new = cover_read_scope(f, ident_ctx, loc_rd);
-            cover_merge_scope(data->root_scope, new);
-         }
-         break;
-      case CTRL_END_OF_FILE:
-         eof = true;
-         break;
-      default:
-         fatal_trace("invalid control word %x in cover db", ctrl);
-      }
-   } while (!eof);
-
-   ident_read_end(ident_ctx);
-   loc_read_end(loc_rd);
-}
-
-unsigned cover_count_items(cover_data_t *data)
-{
-   if (data == NULL)
-      return 0;
-   else
-      return data->next_tag;
-}
-
-void cover_load_spec_file(cover_data_t *data, const char *path)
-{
-   cover_exclude_ctx_t ctx = {};
-
-   ctx.ef = fopen(path, "r");
-   if (ctx.ef == NULL)
-      fatal_errno("failed to open coverage specification file: %s\n", path);
-
-   data->spec = xcalloc(sizeof(cover_spec_t));
-
-   const char *delim = " ";
-   ssize_t read;
-   loc_file_ref_t file_ref = loc_file_ref(path, NULL);
-   int line_num = 1;
-   size_t line_len;
-
-   while ((read = getline(&(ctx.line), &line_len, ctx.ef)) != -1) {
-
-      // Strip comments and newline
-      char *p = ctx.line;
-      for (; *p != '\0' && *p != '#' && *p != '\n'; p++);
-      *p = '\0';
-
-      ctx.loc = get_loc(line_num, 0, line_num, p - ctx.line - 1, file_ref);
-
-      // Parse line as space separated tokens
-      for (char *tok = strtok(ctx.line, delim); tok; tok = strtok(NULL, delim)) {
-         bool include;
-         if (tok[0] == '+')
-            include = true;
-         else if (tok[0] == '-')
-            include = false;
-         else
-            fatal_at(&ctx.loc, "coverage specification command must "
-                               "start with '+' or '-");
-         tok++;
-
-         char_array_t *arr;
-         if (!strcmp(tok, "hierarchy")) {
-            if (include)
-               arr = &(data->spec->hier_include);
-            else
-               arr = &(data->spec->hier_exclude);
-         }
-         else if (!strcmp(tok, "block")) {
-            if (include)
-               arr = &(data->spec->block_include);
-            else
-               arr = &(data->spec->block_exclude);
-         }
-         else if (!strcmp(tok, "fsm_type")) {
-            if (include)
-               arr = &(data->spec->fsm_type_include);
-            else
-               arr = &(data->spec->fsm_type_exclude);
-         }
-         else
-            fatal_at(&ctx.loc, "invalid command: $bold$%s$$", tok);
-
-         char *hier = strtok(NULL, delim);
-         if (!hier)
-            fatal_at(&ctx.loc, "%s name missing", tok);
-
-         APUSH(*arr, xstrdup(hier));
-
-         char *str = AGET(*arr, arr->count - 1);
-         int i = 0;
-         while (str[i]) {
-            str[i] = toupper_iso88591(str[i]);
-            i++;
-         }
-
-         tok = strtok(NULL, delim);
-      }
-      line_num++;
-   }
-
-   fclose(ctx.ef);
-}
-
+//#define COVER_DEBUG_CALLBACK
 
 ///////////////////////////////////////////////////////////////////////////////
 // Runtime handling
@@ -1159,7 +129,6 @@ static inline void cover_toggle_check_0_1_u_z(uint8_t old, uint8_t new,
 #define COVER_TGL_SIGNAL_DETAILS(signal, size)
 #endif
 
-
 #define DEFINE_COVER_TOGGLE_CB(name, check_fnc)                               \
    static void name(uint64_t now, rt_signal_t *s, rt_watch_t *w, void *user)  \
    {                                                                          \
@@ -1177,12 +146,10 @@ static inline void cover_toggle_check_0_1_u_z(uint8_t old, uint8_t new,
       COVER_TGL_SIGNAL_DETAILS(s, s_size)                                     \
    }                                                                          \
 
-
 DEFINE_COVER_TOGGLE_CB(cover_toggle_cb_0_1,     cover_toggle_check_0_1)
 DEFINE_COVER_TOGGLE_CB(cover_toggle_cb_0_1_u,   cover_toggle_check_0_1_u)
 DEFINE_COVER_TOGGLE_CB(cover_toggle_cb_0_1_z,   cover_toggle_check_0_1_z)
 DEFINE_COVER_TOGGLE_CB(cover_toggle_cb_0_1_u_z, cover_toggle_check_0_1_u_z)
-
 
 static bool is_constant_input(rt_signal_t *s)
 {
@@ -1270,1718 +237,4 @@ void x_cover_setup_state_cb(sig_shared_t *ss, int64_t low, int32_t tag)
    *mask |= COV_FLAG_STATE;
 
    model_set_event_cb(m, s, cover_state_cb, (void *)(uintptr_t)(tag - low), false);
-}
-
-///////////////////////////////////////////////////////////////////////////////
-// Exclude file
-///////////////////////////////////////////////////////////////////////////////
-
-static char* cover_bmask_to_bin_list(uint32_t bmask)
-{
-   LOCAL_TEXT_BUF tb = tb_new();
-   bool empty = true;
-   for (int i = 0; i < ARRAY_LEN(bin_map); i++)
-      if (bmask & bin_map[i].flag) {
-         if (!empty)
-            tb_cat(tb, ", ");
-         tb_cat(tb, bin_map[i].name);
-         empty = false;
-      }
-   return tb_claim(tb);
-}
-
-static uint32_t cover_bin_str_to_bmask(cover_exclude_ctx_t *ctx, const char *bin,
-                                       cover_item_t *item)
-{
-   uint32_t allowed = item->flags & COVER_FLAGS_ALL_BINS;
-
-   // If bin is not given, exclude all bins of a item
-   if (!bin)
-      return allowed;
-
-   // Check if bin is valid for given coverage item kind
-   for (int i = 0; i < ARRAY_LEN(bin_map); i++) {
-      if (strcmp(bin, bin_map[i].name))
-         continue;
-      if ((allowed & bin_map[i].flag) != 0)
-         return bin_map[i].flag;
-   }
-
-   char LOCAL *bin_list = cover_bmask_to_bin_list(allowed);
-   fatal_at(&ctx->loc, "invalid bin: $bold$'%s'$$ for %s: '%s'."
-            " Valid bins are: %s", bin, item_kind_str[item->kind],
-            istr(item->hier), bin_list);
-
-   return 0;
-}
-
-static int cover_exclude_item(cover_exclude_ctx_t *ctx, cover_item_t *item,
-                              const char *bin, ident_t hier)
-{
-   int kind = item->kind;
-
-   switch (kind) {
-   case COV_ITEM_STMT:
-      if (bin)
-         fatal_at(&ctx->loc, "statements do not contain bins, but bin '%s' "
-                             "was given for statement: '%s'", bin, istr(hier));
-
-      note_at(&ctx->loc, "excluding statement: '%s'", istr(hier));
-      if (item->data)
-         warn_at(&ctx->loc, "statement: '%s' already covered!", istr(hier));
-
-      item->excl_msk |= 0xFFFFFFFF;
-      return 1;
-
-   case COV_ITEM_BRANCH:
-   case COV_ITEM_TOGGLE:
-   case COV_ITEM_EXPRESSION:
-   {
-      int32_t bmask = cover_bin_str_to_bmask(ctx, bin, item);
-
-      char LOCAL *blist = cover_bmask_to_bin_list(bmask);
-      note_at(&ctx->loc, "excluding %s: '%s' bins: %s",
-               item_kind_str[item->kind], istr(hier), blist);
-
-      uint32_t excl_cov = bmask & item->data;
-      if (excl_cov) {
-         char LOCAL *blist = cover_bmask_to_bin_list(excl_cov);
-         warn_at(&ctx->loc, "%s: '%s' bins: %s already covered!",
-                  item_kind_str[item->kind], istr(hier), blist);
-      }
-
-      item->excl_msk |= bmask;
-      return 1;
-   }
-
-   case COV_ITEM_STATE:
-   {
-      int n_states = item->num;
-      cover_item_t* curr_item = item;
-      for (int i = 0; i < n_states; i++) {
-         ident_t state_name = ident_rfrom(curr_item->hier, '.');
-         if (!bin || !strcmp(istr(state_name), bin)) {
-            if (curr_item->data & COV_FLAG_STATE)
-               warn_at(&ctx->loc, "%s: '%s' of '%s' already covered!",
-                       item_kind_str[item->kind], istr(state_name), istr(hier));
-            curr_item->excl_msk |= COV_FLAG_STATE;
-         }
-         curr_item++;
-      }
-
-      return n_states;
-   }
-
-   default:
-      fatal("Unsupported cover item kind: %d", kind);
-   }
-
-   return 1;
-}
-
-static bool cover_exclude_hier(cover_scope_t *s, cover_exclude_ctx_t *ctx,
-                               const char *excl_hier, const char *bin)
-{
-   bool match = false;
-   int len = strlen(excl_hier);
-   int step;
-
-   for (int i = 0; i < s->items.count; i += step) {
-      cover_item_t *item = AREF(s->items, i);
-      ident_t hier = item->hier;
-
-      // FSM state items contain bin name as part of hierarchy -> Strip it
-      if (item->kind == COV_ITEM_STATE)
-         hier = ident_runtil(hier, '.');
-
-      if (ident_glob(hier, excl_hier, len)) {
-#ifdef COVER_DEBUG_EXCLUDE
-         printf("Applying matching exclude:\n");
-         printf("    Item:        %s\n", istr(hier));
-         printf("    Item flags:  %x\n", item->flags);
-         printf("    Item data:   %x\n", item->data);
-#endif
-         match = true;
-         step = cover_exclude_item(ctx, item, bin, hier);
-      }
-      else
-         step = 1;
-   }
-
-   for (list_iter(cover_scope_t *, it, s->children))
-      match |= cover_exclude_hier(it, ctx, excl_hier, bin);
-
-   return match;
-}
-
-void cover_load_exclude_file(const char *path, cover_data_t *data)
-{
-   cover_exclude_ctx_t ctx = {};
-
-   ctx.ef = fopen(path, "r");
-   if (ctx.ef == NULL)
-      fatal_errno("failed to open exclude file: %s\n", path);
-
-   char *delim = " ";
-   ssize_t read;
-   loc_file_ref_t file_ref = loc_file_ref(path, NULL);
-   int line_num = 1;
-   size_t line_len;
-
-   while ((read = getline(&(ctx.line), &line_len, ctx.ef)) != -1) {
-
-      // Strip comments and newline
-      char *p = ctx.line;
-      for (; *p != '\0' && *p != '#' && *p != '\n'; p++);
-      *p = '\0';
-
-      ctx.loc = get_loc(line_num, 0, line_num, p - ctx.line - 1, file_ref);
-
-      // Parse line as space separated tokens
-      for (char *tok = strtok(ctx.line, delim); tok; tok = strtok(NULL, delim)) {
-         if (!strcmp(tok, "exclude")) {
-            char *excl_hier = strtok(NULL, delim);
-            char *bin = strtok(NULL, delim);
-
-            if (!excl_hier)
-               fatal_at(&ctx.loc, "exclude hierarchy missing!");
-
-            int i = 0;
-            while (excl_hier[i]) {
-               excl_hier[i] = toupper_iso88591(excl_hier[i]);
-               i++;
-            }
-
-            if (!cover_exclude_hier(data->root_scope, &ctx, excl_hier, bin))
-               warn_at(&ctx.loc, "exluded hierarchy does not match any "
-                       "coverage item: '%s'", excl_hier);
-         }
-         else
-            fatal_at(&ctx.loc, "invalid command: $bold$%s$$", tok);
-
-         tok = strtok(NULL, delim);
-      }
-      line_num++;
-   }
-
-   fclose(ctx.ef);
-}
-
-
-///////////////////////////////////////////////////////////////////////////////
-// Report generation
-///////////////////////////////////////////////////////////////////////////////
-
-static void cover_append_line(cover_file_t *f, const char *buf)
-{
-   if (f->n_lines == f->alloc_lines) {
-      f->alloc_lines *= 2;
-      f->lines = xrealloc_array(f->lines, f->alloc_lines, sizeof(cover_line_t));
-   }
-
-   cover_line_t *l = &(f->lines[(f->n_lines)++]);
-   l->text = xstrdup(buf);
-   l->len  = strlen(buf);
-}
-
-static cover_file_t *cover_file(const loc_t *loc)
-{
-   if (loc_invalid_p(loc))
-      return NULL;
-
-   cover_file_t *f;
-   for (f = files; f != NULL; f = f->next) {
-      // Comparing pointers directly here is OK since only one copy
-      // of the file name string will be created by tree_read
-      if (f->name == loc_file_str(loc))
-         return f->valid ? f : NULL;
-   }
-
-   f = xmalloc(sizeof(cover_file_t));
-   f->name        = loc_file_str(loc);
-   f->n_lines     = 0;
-   f->alloc_lines = 1024;
-   f->lines       = xmalloc_array(f->alloc_lines, sizeof(cover_line_t));
-   f->next        = files;
-
-   FILE *fp = fopen(loc_file_str(loc), "r");
-
-   if (fp == NULL) {
-      // Guess the path is relative to the work library
-      char *path LOCAL =
-         xasprintf("%s/../%s", lib_path(lib_work()), loc_file_str(loc));
-      fp = fopen(path, "r");
-   }
-
-   if (fp == NULL) {
-      warnf("failed to open %s for coverage report", loc_file_str(loc));
-      f->valid = false;
-   }
-   else {
-      f->valid = true;
-
-      while (!feof(fp)) {
-         char buf[1024];
-         if (fgets(buf, sizeof(buf), fp) != NULL)
-            cover_append_line(f, buf);
-         else if (ferror(fp))
-            fatal("error reading %s", loc_file_str(loc));
-      }
-
-      fclose(fp);
-   }
-
-   return (files = f);
-}
-
-
-static void cover_print_html_header(FILE *f, cover_report_ctx_t *ctx, bool top,
-                                    cover_scope_t *s, const char *title, ...)
-{
-   fprintf(f, "<!DOCTYPE html>\n"
-              "<html>\n"
-              "  <head>\n"
-              "  <title>");
-
-   va_list ap;
-   va_start(ap, title);
-   vfprintf(f, title, ap);
-   va_end(ap);
-
-   fprintf(f, "</title>\n"
-              "  <style>\n"
-              "   header {\n"
-              "      padding: 30px;\n"
-              "      text-align: center;\n"
-              "      font-size: 35px;\n"
-              "   }\n"
-              "   h2 {\n"
-              "      word-wrap: break-word;\n"
-              "      width:70%%\n"
-              "   }\n"
-              "   h3 {\n"
-              "      word-wrap: break-word;\n"
-              "      margin-bottom: 0px;\n"
-              "   }\n"
-              "   hr {\n"
-              "      border:none;\n"
-              "      height: 2px;\n"
-              "      background: black;\n"
-              "   }\n"
-              "   nav {\n"
-              "      float: left;\n"
-              "      background-color: #ccc;\n"
-              "      width: " SIDEBAR_WIDTH ";\n"
-              "      height: 100%%;\n"
-              "      padding: 10px;\n"
-              "      margin-top: 100px;\n"
-              "      word-wrap: break-word;\n"
-              "     }\n"
-              "   table {\n"
-              "     table-layout: fixed;"
-              "   }\n"
-              "   table, th, td {\n"
-              "     border: 2px solid black;\n"
-              "     border-collapse: collapse;\n"
-              "     word-wrap: break-word;\n"
-              "   }\n"
-              "   .tabcontent {\n"
-              "     display: none;\n"
-              "     padding: 0px 0px;\n"
-              "     border: 2px solid #ccc;\n"
-              "     border-top: none;\n"
-              "     word-wrap: break-word;\n"
-              "   }\n"
-              "   .tab {\n"
-              "     overflow: hidden;\n"
-              "     border: none;\n"
-              "     background-color: none;\n"
-              "     margin-left: " MARGIN_LEFT ";\n"
-              "     margin-top: 10px;\n"
-              "   }\n"
-              "   .tab button.active {\n"
-              "     background-color: #ccc;\n"
-              "   }\n"
-              "   .tab button:hover {\n"
-              "     background-color: #ddd;\n"
-              "   }\n"
-              "   .tab button {\n"
-              "     background-color: inherit;\n"
-              "     float: left;\n"
-              "     margin-left: 20px\n"
-              "     border: 2px solid black;\n"
-              "     cursor: pointer;\n"
-              "     padding: 14px 16px;\n"
-              "     font-size: 17px;\n"
-              "   }\n"
-              "   .cbg {\n"
-              "     background-color: #bbbbbb;\n"
-              "   }\n"
-              "   .cbt {\n"
-              "     margin-top: 8px;\n"
-              "   }\n"
-              "   .cbt th {\n"
-              "     background-color: #bbbbbb;\n"
-              "     text-align: center;\n"
-              "    }\n"
-              "   .cbt td, .cbt th {\n"
-              "     width:50px;\n"
-              "     text-align: center;\n"
-              "    }\n"
-              "   .cbt td + td, .cbt th + th { width:150px; }\n"
-              "   .cbt td + td + td, .cbt th + th + th { width:150px; }\n"
-              "   .cbt td + td + td + td, .cbt th + th + th + th { width:150px; }\n"
-              "  </style>\n"
-              "  </head>\n"
-              "  <section>\n\n");
-
-   if (!top) {
-      fprintf(f, "<nav>\n"
-                 "<b>Hierarchy:</b><br>\n");
-      int offset = 0;
-
-      ident_t full_hier = s->hier;
-      ident_t curr_id;
-      ident_t curr_hier = NULL;
-      const char *link = "../index";
-      do {
-         curr_id = ident_walk_selected(&full_hier);
-         curr_hier = ident_prefix(curr_hier, curr_id, '.');
-         if (offset > 0)
-            link = istr(curr_hier);
-         if (curr_id)
-            fprintf(f, "<p style=\"margin-left: %dpx\"><a href=%s.html>%s</a></p>\n",
-                        offset * 10, link, istr(curr_id));
-         offset++;
-      } while (curr_id != NULL);
-      fprintf(f, "</nav>\n\n");
-   }
-
-   fprintf(f, "<header>");
-   va_start(ap, title);
-   vfprintf(f, title, ap);
-   va_end(ap);
-   fprintf(f, "</header>\n\n");
-
-   fprintf(f, "<h2 style=\"margin-left: " MARGIN_LEFT ";\">\n");
-   if (!top)
-      fprintf(f, "   Instance:&nbsp;%s\n", istr(s->hier));
-   else
-      fprintf(f, "   Instance:");
-   fprintf(f, "</h2>\n\n");
-
-   cover_file_t *src = cover_file(&(s->loc));
-   fprintf(f, "<h2 style=\"margin-left: " MARGIN_LEFT ";\">\n");
-   if (!top)
-      fprintf(f, "   File:&nbsp; <a href=\"../../%s\">../../%s</a>\n",
-               src->name, src->name);
-   else
-      fprintf(f, "   File:");
-   fprintf(f, "</h2>\n\n");
-}
-
-static void cover_print_percents_cell(FILE *f, unsigned hit, unsigned total)
-{
-   if (total > 0) {
-      float perc = ((float) hit / (float) total) * 100;
-      char color[8];
-      if (hit == total)
-         checked_sprintf(color, sizeof(color), "#00cc00");
-      else if (perc > 90)
-         checked_sprintf(color, sizeof(color), "#e6e600");
-      else if (perc > 80)
-         checked_sprintf(color, sizeof(color), "#ff9900");
-      else
-         checked_sprintf(color, sizeof(color), "#ff0000");
-
-      fprintf(f, "    <td bgcolor=%s>%.1f %% (%d/%d)</td>\n",
-              color, perc, hit, total);
-      return;
-   }
-
-   fprintf(f, "    <td bgcolor=#aaaaaa>N.A.</td>\n");
-}
-
-static void cover_print_hierarchy_header(FILE *f)
-{
-   fprintf(f, "<table style=\"width:70%%;margin-left:" MARGIN_LEFT ";margin-right:auto;\"> \n"
-              "  <tr>\n"
-              "    <th class=\"cbg\" style=\"width:30%%\">Instance</th>\n"
-              "    <th class=\"cbg\" style=\"width:8%%\">Statement</th>\n"
-              "    <th class=\"cbg\" style=\"width:8%%\">Branch</th>\n"
-              "    <th class=\"cbg\" style=\"width:8%%\">Toggle</th>\n"
-              "    <th class=\"cbg\" style=\"width:8%%\">Expression</th>\n"
-              "    <th class=\"cbg\" style=\"width:8%%\">FSM state</th>\n"
-              "    <th class=\"cbg\" style=\"width:8%%\">Average</th>\n"
-              "  </tr>\n");
-}
-
-static void cover_print_hierarchy_footer(FILE *f)
-{
-   fprintf(f, "</table>\n\n");
-}
-
-static void cover_print_timestamp(FILE *f)
-{
-   time_t t;
-   time(&t);
-
-   fprintf(f, "  </section>\n");
-
-   fprintf(f, "<footer>");
-   fprintf(f, "   <p> NVC version: %s </p>\n", PACKAGE_VERSION);
-   fprintf(f, "   <p> Generated on: %s </p>\n", ctime(&t));
-   fprintf(f,  "</footer>");
-}
-
-static void cover_print_hierarchy_summary(FILE *f, cover_report_ctx_t *ctx, ident_t hier,
-                                          bool top, bool full_hier, bool flat)
-{
-   ident_t print_hier = (full_hier) ? hier : ident_rfrom(hier, '.');
-   cover_stats_t *stats;
-
-   if (flat)
-      stats = &(ctx->flat_stats);
-   else
-      stats = &(ctx->nested_stats);
-
-   fprintf(f, "  <tr>\n"
-              "    <td><a href=\"%s%s.html\">%s</a></td>\n",
-              top ? "hier/" : "", istr(hier), istr(print_hier));
-
-   cover_print_percents_cell(f, stats->hit_stmts, stats->total_stmts);
-   cover_print_percents_cell(f, stats->hit_branches, stats->total_branches);
-   cover_print_percents_cell(f, stats->hit_toggles, stats->total_toggles);
-   cover_print_percents_cell(f, stats->hit_expressions, stats->total_expressions);
-   cover_print_percents_cell(f, stats->hit_states, stats->total_states);
-
-   int avg_total = stats->total_stmts + stats->total_branches +
-                   stats->total_toggles + stats->total_expressions +
-                   stats->total_states;
-   int avg_hit = stats->hit_stmts + stats->hit_branches +
-                 stats->hit_toggles + stats->hit_expressions +
-                 stats->hit_states;
-
-   cover_print_percents_cell(f, avg_hit, avg_total);
-
-   fprintf(f, "  </tr>\n");
-
-   float perc_stmt = 0.0f;
-   float perc_branch = 0.0f;
-   float perc_toggle = 0.0f;
-   float perc_expr = 0.0f;
-   float perc_state = 0.0f;
-
-   if (stats->total_stmts > 0)
-      perc_stmt = 100.0 * ((float)stats->hit_stmts) / stats->total_stmts;
-   if (stats->total_branches > 0)
-      perc_branch = 100.0 * ((float)stats->hit_branches) / stats->total_branches;
-   if (stats->total_toggles > 0)
-      perc_toggle = 100.0 * ((float)stats->hit_toggles) / stats->total_toggles;
-   if (stats->total_expressions > 0)
-      perc_expr = 100.0 * ((float)stats->hit_expressions) / stats->total_expressions;
-   if (stats->total_states > 0)
-      perc_state = 100.0 * ((float)stats->hit_states) / stats->total_states;
-
-   if (top) {
-      notef("code coverage results for: %s", istr(hier));
-
-      if (perc_stmt > 0)
-         notef("     statement:     %.1f %% (%d/%d)", perc_stmt,
-               stats->hit_stmts, stats->total_stmts);
-      else
-         notef("     statement:     N.A.");
-
-      if (perc_branch > 0)
-         notef("     branch:        %.1f %% (%d/%d)", perc_branch,
-               stats->hit_branches, stats->total_branches);
-      else
-         notef("     branch:        N.A.");
-
-      if (perc_toggle > 0)
-         notef("     toggle:        %.1f %% (%d/%d)", perc_toggle,
-               stats->hit_toggles, stats->total_toggles);
-      else
-         notef("     toggle:        N.A.");
-
-      if (perc_expr > 0)
-         notef("     expression:    %.1f %% (%d/%d)", perc_expr,
-               stats->hit_expressions, stats->total_expressions);
-      else
-         notef("     expression:    N.A.");
-
-      if (perc_state > 0)
-         notef("     FSM state:     %.1f %% (%d/%d)", perc_state,
-               stats->hit_states, stats->total_states);
-      else
-         notef("     FSM state:     N.A.");
-   }
-   else if (opt_get_int(OPT_VERBOSE) && !flat) {
-
-      cover_rpt_buf_t *new = xcalloc(sizeof(cover_rpt_buf_t));
-      new->tb = tb_new();
-      new->prev = ctx->data->rpt_buf;
-      ctx->data->rpt_buf = new;
-
-      tb_printf(new->tb,
-         "%*s %-*s %10.1f %% (%d/%d)  %10.1f %% (%d/%d) %10.1f %% (%d/%d) "
-         "%10.1f %% (%d/%d) %10.1f %% (%d/%d)",
-         ctx->lvl, "", 50-ctx->lvl, istr(ident_rfrom(hier, '.')),
-         perc_stmt, stats->hit_stmts, stats->total_stmts,
-         perc_branch, stats->hit_branches, stats->total_branches,
-         perc_toggle, stats->hit_toggles, stats->total_toggles,
-         perc_expr, stats->hit_expressions, stats->total_expressions,
-         perc_state, stats->hit_states, stats->total_states);
-   }
-}
-
-static inline void cover_print_char(FILE *f, char c)
-{
-   if (c == '\n' || c == ' ')
-      fprintf(f, "&nbsp");
-   else
-      fprintf(f, "%c", c);
-}
-
-static void cover_print_single_code_line(FILE *f, loc_t loc, cover_line_t *line)
-{
-   assert(loc.line_delta == 0);
-
-   int curr_char = 0;
-   while (curr_char < strlen(line->text)) {
-
-      // Highlight code location
-      if (curr_char == loc.first_column)
-         fprintf(f, "<code class=\"cbg\">");
-
-      cover_print_char(f, line->text[curr_char]);
-
-      // Finish code highlight
-      if (curr_char == (loc.first_column + loc.column_delta))
-         fprintf(f, "</code>");
-
-      curr_char++;
-   }
-}
-
-static void cover_print_lhs_rhs_arrows(FILE *f, cover_pair_t *pair)
-{
-   int last = strlen(pair->line->text);
-   int curr = 0;
-
-   // Offset by line number width
-   int digits = pair->item->loc.first_line;
-   do {
-      digits /= 10;
-      fprintf(f, "&nbsp");
-   } while (digits > 0);
-   fprintf(f, "&nbsp");
-
-   loc_t *loc_lhs = &(pair->item->loc_lhs);
-   loc_t *loc_rhs = &(pair->item->loc_rhs);
-
-   int lhs_beg = loc_lhs->first_column;
-   int rhs_beg = loc_rhs->first_column;
-
-   int lhs_end = lhs_beg + loc_lhs->column_delta;
-   int rhs_end = rhs_beg + loc_rhs->column_delta;
-
-   int lhs_mid = (lhs_end + lhs_beg) / 2;
-   int rhs_mid = (rhs_end + rhs_beg) / 2;
-
-   while (curr < last) {
-      if (curr == lhs_mid - 1)
-         fprintf(f, "L");
-      else if (curr == lhs_mid)
-         fprintf(f, "H");
-      else if (curr == lhs_mid + 1)
-         fprintf(f, "S");
-
-      else if (curr == lhs_beg)
-         fprintf(f, "<");
-      else if (curr > lhs_beg && curr < lhs_end)
-         fprintf(f, "-");
-      else if (curr == (lhs_beg + loc_lhs->column_delta))
-         fprintf(f, ">");
-
-      else if (curr == rhs_mid - 1)
-         fprintf(f, "R");
-      else if (curr == rhs_mid)
-         fprintf(f, "H");
-      else if (curr == rhs_mid + 1)
-         fprintf(f, "S");
-
-      else if (curr == rhs_beg)
-         fprintf(f, "<");
-      else if (curr > rhs_beg && curr < rhs_end)
-         fprintf(f, "-");
-      else if (curr == (rhs_beg + loc_rhs->column_delta))
-         fprintf(f, ">");
-
-      else
-         fprintf(f, "&nbsp");
-
-      curr++;
-   }
-}
-
-static void cover_print_item_title(FILE *f, cover_pair_t *pair)
-{
-   tree_kind_t kind = pair->item->tree_kind;
-
-   fprintf(f, "<h3>");
-
-   switch (pair->item->kind) {
-   case COV_ITEM_STMT:
-      fprintf(f, "\"%s\" ", tree_kind_str(kind));
-      fprintf(f, "statement:");
-      break;
-   case COV_ITEM_BRANCH:
-      switch (kind) {
-      case T_COND_STMT:
-      case T_COND_ASSIGN:
-         fprintf(f, "\"if\" / \"when\" / \"else\" condition:");
-         break;
-      case T_ASSOC:
-         fprintf(f, "\"case\" / \"with\" / \"select\" choice:");
-         break;
-      case T_WHILE:
-      case T_EXIT:
-      case T_NEXT:
-         fprintf(f, "Loop control condition:");
-         break;
-      default:
-         fprintf(f, "Condition:");
-         break;
-      };
-      break;
-   case COV_ITEM_EXPRESSION:
-      fprintf(f, "%s expression", istr(pair->item->func_name));
-      break;
-   case COV_ITEM_STATE:
-      fprintf(f, "\"%s\" FSM with total %d states", istr(pair->item->func_name),
-              pair->item->num);
-      break;
-   default:
-      break;
-   }
-   fprintf(f, "</h3>");
-}
-
-static void cover_print_code_loc(FILE *f, cover_pair_t *pair)
-{
-   loc_t loc = pair->item->loc;
-   cover_line_t *curr_line = pair->line;
-   cover_line_t *last_line = pair->line + loc.line_delta;
-
-   cover_print_item_title(f, pair);
-
-   if (loc.line_delta == 0) {
-      fprintf(f, "<code>");
-      fprintf(f, "%d:", loc.first_line);
-
-      cover_print_single_code_line(f, loc, curr_line);
-      if (pair->item->flags & COVER_FLAGS_LHS_RHS_BINS) {
-         fprintf(f, "<br>");
-         cover_print_lhs_rhs_arrows(f, pair);
-      }
-      fprintf(f, "</code>");
-   }
-   else {
-      fprintf(f, "<code>");
-
-      do {
-         // Shorten code samples longer than 5 lines
-         if (loc.line_delta > 5 &&
-             curr_line == pair->line + 2) {
-            fprintf(f, "...<br>");
-            curr_line = last_line - 1;
-            continue;
-         }
-         else
-            fprintf(f, "%lu:", loc.first_line + (curr_line - pair->line));
-
-         int curr_char = 0;
-         while (curr_char < curr_line->len) {
-            cover_print_char(f, curr_line->text[curr_char]);
-            curr_char++;
-         }
-
-         fprintf(f, "<br>");
-         curr_line++;
-
-      } while (curr_line <= last_line);
-
-      fprintf(f, "</code>");
-   }
-}
-
-static void cover_print_get_exclude_button(FILE *f, cover_item_t *item,
-                                           uint32_t flag, bool add_td)
-{
-
-   const char *bin_name = "";
-   if (flag) {
-      for (int i = 0; i < ARRAY_LEN(bin_map); i++) {
-         if (bin_map[i].flag == flag)
-            bin_name = bin_map[i].name;
-      }
-   }
-
-   if (add_td)
-      fprintf(f, "<td>");
-
-   ident_t hier = item->hier;
-
-   // State coverage contains bin name (state name) appended to hierarchical path
-   if (item->kind == COV_ITEM_STATE) {
-      bin_name = istr(ident_rfrom(hier, '.'));
-      hier = ident_runtil(hier, '.');
-   }
-
-   fprintf(f, "<button onclick=\"GetExclude('exclude %s %s')\" %s>"
-              "Copy %sto Clipoard</button>", istr(hier), bin_name,
-           (item->kind == COV_ITEM_STMT) ? "style=\"float: right;\"" : "",
-           (item->kind == COV_ITEM_STMT) ? "Exclude Command " : "");
-
-   if (add_td)
-      fprintf(f, "</td>");
-}
-
-static void cover_print_bin(FILE *f, cover_pair_t *pair, uint32_t flag,
-                            cov_pair_kind_t pkind, int cols, ...)
-{
-   va_list argp;
-   va_start(argp, cols);
-
-   if (pair->flags & flag) {
-      fprintf(f, "<tr><td><b>Bin</b></td>");
-
-      for (int i = 0; i < cols; i++) {
-         const char *val = va_arg(argp, const char *);
-         fprintf(f, "<td>%s</td>", val);
-      }
-
-      if (pkind == PAIR_UNCOVERED)
-         cover_print_get_exclude_button(f, pair->item, flag, true);
-
-      if (pkind == PAIR_EXCLUDED) {
-         const char *er = (flag & pair->item->unrc_msk) ?
-            "Unreachable" : "Exclude file";
-         fprintf(f, "<td>%s</td>", er);
-      }
-
-      fprintf(f, "</tr>");
-   }
-}
-
-static void cover_print_bin_header(FILE *f, cov_pair_kind_t pkind, int cols, ...)
-{
-   va_list argp;
-   va_start(argp, cols);
-
-   fprintf(f, "<tr><th></th>");
-
-   for (int i = 0; i < cols; i++) {
-      const char *val = va_arg(argp, const char *);
-      fprintf(f, "<th>%s</th>", val);
-   }
-
-   if (pkind == PAIR_UNCOVERED)
-      fprintf(f, "<th>Exclude Command</th>");
-
-   if (pkind == PAIR_EXCLUDED)
-      fprintf(f, "<th>Excluded due to</th>");
-
-   fprintf(f, "</tr>");
-}
-
-static void cover_print_bins(FILE *f, cover_pair_t *pair, cov_pair_kind_t pkind)
-{
-   loc_t loc = pair->item->loc;
-
-   fprintf(f, "<br><table class=\"cbt\">");
-
-   switch (pair->item->kind) {
-   case COV_ITEM_BRANCH:
-      cover_print_bin_header(f, pkind, 1, (pair->flags & COV_FLAG_CHOICE) ?
-                              "Choice of" : "Evaluated to");
-      cover_print_bin(f, pair, COV_FLAG_TRUE, pkind, 1, "True");
-      cover_print_bin(f, pair, COV_FLAG_FALSE, pkind, 1, "False");
-
-      if (pair->flags & COV_FLAG_CHOICE) {
-         int curr = loc.first_column;
-         int last = (loc.line_delta) ? strlen(pair->line->text) :
-                                       loc.column_delta + curr;
-
-         LOCAL_TEXT_BUF tb = tb_new();
-         tb_printf(tb, "<code>");
-         while (curr <= last) {
-            tb_printf(tb, "%c", pair->line->text[curr]);
-            curr++;
-         }
-         tb_printf(tb, "</code>");
-
-         cover_print_bin(f, pair, COV_FLAG_CHOICE, pkind, 1, tb_get(tb));
-      }
-      break;
-
-   case COV_ITEM_TOGGLE:
-      cover_print_bin_header(f, pkind, 2, "From", "To");
-      cover_print_bin(f, pair, COV_FLAG_TOGGLE_TO_1, pkind, 2, "0", "1");
-      cover_print_bin(f, pair, COV_FLAG_TOGGLE_TO_0, pkind, 2, "1", "0");
-      break;
-
-   case COV_ITEM_EXPRESSION:
-      {
-      char *t_str = (pair->item->flags & COV_FLAG_EXPR_STD_LOGIC) ? "'1'" : "True";
-      char *f_str = (pair->item->flags & COV_FLAG_EXPR_STD_LOGIC) ? "'0'" : "False";
-
-      if ((pair->flags & COV_FLAG_TRUE) || (pair->flags & COV_FLAG_FALSE)) {
-         cover_print_bin_header(f, pkind, 1, "Evaluated to");
-         cover_print_bin(f, pair, COV_FLAG_TRUE, pkind, 1, t_str);
-         cover_print_bin(f, pair, COV_FLAG_FALSE, pkind, 1, f_str);
-      }
-
-      if (pair->flags & COV_FLAG_00 || pair->flags & COV_FLAG_01 ||
-          pair->flags & COV_FLAG_10 || pair->flags & COV_FLAG_11) {
-         cover_print_bin_header(f, pkind, 2, "LHS", "RHS");
-
-         cover_print_bin(f, pair, COV_FLAG_00, pkind, 2, f_str, f_str);
-         cover_print_bin(f, pair, COV_FLAG_01, pkind, 2, f_str, t_str);
-         cover_print_bin(f, pair, COV_FLAG_10, pkind, 2, t_str, f_str);
-         cover_print_bin(f, pair, COV_FLAG_11, pkind, 2, t_str, t_str);
-      }
-      }
-      break;
-   default:
-      fatal("unsupported type of code coverage: %d !", pair->item->kind);
-   }
-   fprintf(f, "</table>");
-}
-
-
-static int cover_print_fsm_table(FILE *f, cover_pair_t *pair, cov_pair_kind_t pkind,
-                                 cover_pair_t *last)
-{
-   // All pairs of a single FSM are consequent. Group them together and print them
-   // in a single table.
-   ident_t first_prefix = ident_runtil(pair->item->hier, '.');
-
-   fprintf(f, "<br><table class=\"cbt\">");
-
-   cover_print_bin_header(f, pkind, 1, "State");
-
-   int n_pairs = 0;
-   do {
-      ident_t state_name = ident_rfrom(pair->item->hier, '.');
-      cover_print_bin(f, pair, COV_FLAG_STATE, pkind, 1, istr(state_name));
-      n_pairs++;
-
-      // End of the chain was hit
-      if (pair == last)
-         break;
-      pair++;
-
-      // A item with different prefix -> Not the same FSM.
-      ident_t curr_prefix = ident_runtil(pair->item->hier, '.');
-      if (!ident_starts_with(curr_prefix, first_prefix))
-         break;
-
-   } while (1);
-
-   fprintf(f, "</table>");
-
-   return n_pairs;
-}
-
-
-static void cover_print_pairs(FILE *f, cover_pair_t *first, cov_pair_kind_t pkind,
-                              int pair_cnt)
-{
-   if (pair_cnt == 0)
-      return;
-
-   cover_pair_t *last = first + pair_cnt - 1;
-   cover_pair_t *curr = first;
-   int step;
-
-   do {
-      step = 1;
-      fprintf(f, "    <p>");
-
-      switch (curr->item->kind) {
-      case COV_ITEM_STMT:
-         if (pkind == PAIR_UNCOVERED)
-            cover_print_get_exclude_button(f, curr->item, 0, false);
-         if (pkind == PAIR_EXCLUDED) {
-            // No un-reachability on statements so far
-            assert(curr->item->unrc_msk == 0);
-            fprintf(f, "<div style=\"float: right\"><b>Excluded due to:</b> Exclude file</div>");
-         }
-
-         cover_print_code_loc(f, curr);
-         break;
-
-      case COV_ITEM_BRANCH:
-         cover_print_code_loc(f, curr);
-         cover_print_bins(f, curr, pkind);
-         break;
-
-      case COV_ITEM_TOGGLE:
-         if (curr->item->flags & COV_FLAG_TOGGLE_SIGNAL)
-            fprintf(f, "<h3>Signal:</h3>");
-         else if (curr->item->flags & COV_FLAG_TOGGLE_PORT)
-            fprintf(f, "<h3>Port:</h3>");
-
-         const char *sig_name = istr(curr->item->hier);
-         sig_name += curr->item->num;
-         fprintf(f, "&nbsp;<code>%s</code>", sig_name);
-
-         cover_print_bins(f, curr, pkind);
-         break;
-
-      case COV_ITEM_EXPRESSION:
-         cover_print_code_loc(f, curr);
-         cover_print_bins(f, curr, pkind);
-         break;
-
-      case COV_ITEM_STATE:
-         cover_print_code_loc(f, curr);
-         step = cover_print_fsm_table(f, curr, pkind, last);
-         break;
-
-      default:
-         fatal("unsupported type of code coverage: %d !", curr->item->kind);
-      }
-
-      fprintf(f, "<hr>");
-      fprintf(f, "</p>\n");
-
-      curr += step;
-
-   } while (curr <= last);
-}
-
-static void cover_print_chain(FILE *f, cover_data_t *data, cover_chain_t *chn,
-                              cover_item_kind_t kind)
-{
-   // HTML TAB
-   fprintf(f, "<div id=\"");
-   if (kind == COV_ITEM_STMT)
-      fprintf(f, "Statement");
-   else if (kind == COV_ITEM_BRANCH)
-      fprintf(f, "Branch");
-   else if (kind == COV_ITEM_TOGGLE)
-      fprintf(f, "Toggle");
-   else if (kind == COV_ITEM_EXPRESSION)
-      fprintf(f, "Expression");
-   else if (kind == COV_ITEM_STATE)
-      fprintf(f, "FSM state");
-
-   fprintf(f, "\" class=\"tabcontent\" style=\"width:68.5%%;margin-left:" MARGIN_LEFT "; "
-                          "margin-right:auto; margin-top:10px; border: 2px solid black;\">\n");
-
-   for (cov_pair_kind_t pkind = PAIR_UNCOVERED; pkind < PAIR_LAST; pkind++) {
-      int n;
-      cover_pair_t *first_pair;
-
-      if (pkind == PAIR_UNCOVERED) {
-         if (cover_enabled(data, COVER_MASK_DONT_PRINT_UNCOVERED))
-            continue;
-         first_pair = chn->miss;
-         n = chn->n_miss;
-      }
-      else if (pkind == PAIR_EXCLUDED) {
-         if (cover_enabled(data, COVER_MASK_DONT_PRINT_EXCLUDED))
-            continue;
-         first_pair = chn->excl;
-         n = chn->n_excl;
-      }
-      else {
-         if (cover_enabled(data, COVER_MASK_DONT_PRINT_COVERED))
-            continue;
-         first_pair = chn->hits;
-         n = chn->n_hits;
-      }
-
-      fprintf(f, "  <section style=\"background-color:");
-      if (pkind == PAIR_UNCOVERED)
-         fprintf(f, "#ffcccc;\">\n");
-      else if (pkind == PAIR_EXCLUDED)
-         fprintf(f, "#d6eaf8;\">\n");
-      else
-         fprintf(f, "#ccffcc;\">\n");
-
-      fprintf(f, " <h2>");
-      if (pkind == PAIR_UNCOVERED)
-         fprintf(f, "Uncovered ");
-      else if (pkind == PAIR_EXCLUDED)
-         fprintf(f, "Excluded ");
-      else
-         fprintf(f, "Covered ");
-
-      if (kind == COV_ITEM_STMT)
-         fprintf(f, "statements:");
-      else if (kind == COV_ITEM_BRANCH)
-         fprintf(f, "branches:");
-      else if (kind == COV_ITEM_TOGGLE)
-         fprintf(f, "toggles:");
-      else if (kind == COV_ITEM_EXPRESSION)
-         fprintf(f, "expressions:");
-      else if (kind == COV_ITEM_STATE)
-         fprintf(f, "FSM states:");
-      fprintf(f, "</h2>\n");
-
-      fprintf(f, "  <section style=\"padding:0px 10px;\">\n");
-      cover_print_pairs(f, first_pair, pkind, n);
-      fprintf(f, "  </section>\n\n");
-   }
-
-   fprintf(f, "</div>\n");
-}
-
-static void cover_print_hierarchy_guts(FILE *f, cover_report_ctx_t *ctx)
-{
-   fprintf(f, "<div class=\"tab\">"
-              "   <button class=\"tablinks\" onclick=\"selectCoverage(event, 'Statement')\" id=\"defaultOpen\">Statement</button>\n"
-              "   <button class=\"tablinks\" style=\"margin-left:10px;\" onclick=\"selectCoverage(event, 'Branch')\">Branch</button>\n"
-              "   <button class=\"tablinks\" style=\"margin-left:10px;\" onclick=\"selectCoverage(event, 'Toggle')\">Toggle</button>\n"
-              "   <button class=\"tablinks\" style=\"margin-left:10px;\" onclick=\"selectCoverage(event, 'Expression')\">Expression</button>\n"
-              "   <button class=\"tablinks\" style=\"margin-left:10px;\" onclick=\"selectCoverage(event, 'FSM state')\">FSM state</button>\n"
-              "</div>\n\n");
-
-   cover_print_chain(f, ctx->data, &(ctx->ch_stmt), COV_ITEM_STMT);
-   cover_print_chain(f, ctx->data, &(ctx->ch_branch), COV_ITEM_BRANCH);
-   cover_print_chain(f, ctx->data, &(ctx->ch_toggle), COV_ITEM_TOGGLE);
-   cover_print_chain(f, ctx->data, &(ctx->ch_expression), COV_ITEM_EXPRESSION);
-   cover_print_chain(f, ctx->data, &(ctx->ch_state), COV_ITEM_STATE);
-
-   fprintf(f, "<script>\n"
-              "   document.getElementById(\"defaultOpen\").click();"
-              "   function selectCoverage(evt, coverageType) {\n"
-              "      var i, tabcontent, tablinks;\n"
-              "      tabcontent = document.getElementsByClassName(\"tabcontent\");\n"
-              "      for (i = 0; i < tabcontent.length; i++) {\n"
-              "         tabcontent[i].style.display = \"none\";\n"
-              "      }\n"
-              "      tablinks = document.getElementsByClassName(\"tablinks\");\n"
-              "      for (i = 0; i < tablinks.length; i++) {\n"
-              "         tablinks[i].className = tablinks[i].className.replace(\" active\", \"\");\n"
-              "      }\n"
-              "      document.getElementById(coverageType).style.display = \"block\";\n"
-              "      evt.currentTarget.className += \" active\";\n"
-              "   }\n"
-              "   function GetExclude(excludeCmd) {\n"
-              "      navigator.clipboard.writeText(excludeCmd);\n"
-              "   }\n"
-              "</script>\n");
-}
-
-static int cover_append_to_chain(cover_chain_t *chain, cover_item_t *item,
-                                 cover_line_t *line, unsigned hits,
-                                 unsigned misses, unsigned excludes,
-                                 int limit)
-{
-   int rv = 0;
-   if (hits) {
-      if (chain->n_hits <= limit) {
-         if (chain->n_hits == chain->alloc_hits) {
-            chain->alloc_hits *= 2;
-            chain->hits = xrealloc_array(chain->hits, chain->alloc_hits,
-                                       sizeof(cover_pair_t));
-         }
-         chain->hits[chain->n_hits].item = item;
-         chain->hits[chain->n_hits].line = line;
-         chain->hits[chain->n_hits].flags = hits;
-         chain->n_hits++;
-      }
-      else
-         rv++;
-   }
-
-   if (misses) {
-      if (chain->n_miss <= limit) {
-         if (chain->n_miss == chain->alloc_miss) {
-            chain->alloc_miss *= 2;
-            chain->miss = xrealloc_array(chain->miss, chain->alloc_miss,
-                                       sizeof(cover_pair_t));
-         }
-         chain->miss[chain->n_miss].item = item;
-         chain->miss[chain->n_miss].line = line;
-         chain->miss[chain->n_miss].flags = misses;
-         chain->n_miss++;
-      }
-      else
-         rv++;
-   }
-
-   if (excludes) {
-      if (chain->n_excl <= limit) {
-         if (chain->n_excl == chain->alloc_excl) {
-            chain->alloc_excl *= 2;
-            chain->excl = xrealloc_array(chain->excl, chain->alloc_excl,
-                                       sizeof(cover_pair_t));
-         }
-         chain->excl[chain->n_excl].item = item;
-         chain->excl[chain->n_excl].line = line;
-         chain->excl[chain->n_excl].flags = excludes;
-         chain->n_excl++;
-      }
-      else
-         rv++;
-   }
-
-   return rv;
-}
-
-static bool cover_bin_unreachable(cover_report_ctx_t *ctx, cover_item_t *item,
-                                  unsigned flag)
-{
-   if ((ctx->data->mask & COVER_MASK_EXCLUDE_UNREACHABLE) == 0)
-      return false;
-
-   // Toggle items remember unreachability in run-time data. Must check item kind
-   // not to get false unreachability on statement items. Excludes both bins
-   // automatically!
-   if (item->kind == COV_ITEM_TOGGLE && ((item->data & COV_FLAG_CONST_DRIVEN) != 0))
-      return true;
-
-   // Expression items remembed unreachability via dedicated mask
-   if (item->kind == COV_ITEM_EXPRESSION && ((flag & item->unrc_msk) != 0))
-      return true;
-
-   return false;
-}
-
-static void cover_item_to_chain(cover_report_ctx_t *ctx, cover_item_t *item,
-                                cover_flags_t flag, unsigned *hits,
-                                unsigned *misses, unsigned *excludes)
-{
-   unsigned *flat_total;
-   unsigned *nested_total;
-   unsigned *flat_hits;
-   unsigned *nested_hits;
-
-   switch (item->kind) {
-   case COV_ITEM_BRANCH:
-      flat_total = &(ctx->flat_stats.total_branches);
-      nested_total = &(ctx->nested_stats.total_branches);
-      flat_hits = &(ctx->flat_stats.hit_branches);
-      nested_hits = &(ctx->nested_stats.hit_branches);
-      break;
-   case COV_ITEM_TOGGLE:
-      flat_total = &(ctx->flat_stats.total_toggles);
-      nested_total = &(ctx->nested_stats.total_toggles);
-      flat_hits = &(ctx->flat_stats.hit_toggles);
-      nested_hits = &(ctx->nested_stats.hit_toggles);
-      break;
-   case COV_ITEM_EXPRESSION:
-      flat_total = &(ctx->flat_stats.total_expressions);
-      nested_total = &(ctx->nested_stats.total_expressions);
-      flat_hits = &(ctx->flat_stats.hit_expressions);
-      nested_hits = &(ctx->nested_stats.hit_expressions);
-      break;
-   case COV_ITEM_STATE:
-      flat_total = &(ctx->flat_stats.total_states);
-      nested_total = &(ctx->nested_stats.total_states);
-      flat_hits = &(ctx->flat_stats.hit_states);
-      nested_hits = &(ctx->nested_stats.hit_states);
-      break;
-   default:
-      fatal("unsupported type of code coverage: %d !", item->kind);
-   }
-
-   (*flat_total)++;
-   (*nested_total)++;
-
-   if (item->data & flag) {
-      (*flat_hits)++;
-      (*nested_hits)++;
-      (*hits) |= flag;
-   }
-   else if ((item->excl_msk & flag) || cover_bin_unreachable(ctx, item, flag)) {
-      (*flat_hits)++;
-      (*nested_hits)++;
-      (*excludes) |= flag;
-   }
-   else
-      (*misses) |= flag;
-
-}
-
-
-static void cover_report_scope(cover_report_ctx_t *ctx,
-                               cover_scope_t *s, const char *dir,
-                               FILE *summf, int *skipped)
-{
-   for (int i = 0; i < s->items.count; i++) {
-      cover_item_t *item = &(s->items.items[i]);
-
-      cover_file_t *f_src = cover_file(&(item->loc));
-      if (f_src == NULL)
-         continue;
-
-      cover_line_t *line = &(f_src->lines[item->loc.first_line-1]);
-
-      unsigned hits = 0;
-      unsigned misses = 0;
-      unsigned excludes = 0;
-      int limit = ctx->data->report_item_limit;
-
-      switch (item->kind) {
-      case COV_ITEM_STMT:
-         (ctx->flat_stats.total_stmts)++;
-         (ctx->nested_stats.total_stmts)++;
-
-         hits = (item->data != 0);
-         misses = (item->data == 0) && (item->excl_msk == 0);
-         excludes = (item->data == 0) && (item->excl_msk != 0);
-
-         if (hits | excludes) {
-            (ctx->flat_stats.hit_stmts)++;
-            (ctx->nested_stats.hit_stmts)++;
-         }
-         *skipped += cover_append_to_chain(&(ctx->ch_stmt), item, line,
-                                           hits, misses, excludes, limit);
-         break;
-
-      case COV_ITEM_BRANCH:
-         // If/else, when else
-         if (item->flags & COV_FLAG_TRUE && item->flags & COV_FLAG_FALSE) {
-            cover_item_to_chain(ctx, item, COV_FLAG_TRUE,
-                                &hits, &misses, &excludes);
-            cover_item_to_chain(ctx, item, COV_FLAG_FALSE,
-                                &hits, &misses, &excludes);
-            *skipped += cover_append_to_chain(&(ctx->ch_branch), item, line,
-                                              hits, misses, excludes, limit);
-         }
-
-         // Case, with select
-         if (item->flags & COV_FLAG_CHOICE) {
-            cover_item_to_chain(ctx, item, COV_FLAG_CHOICE,
-                                &hits, &misses, &excludes);
-            *skipped += cover_append_to_chain(&(ctx->ch_branch), item, line,
-                                              hits, misses, excludes, limit);
-         }
-         break;
-
-      case COV_ITEM_TOGGLE:
-         cover_item_to_chain(ctx, item, COV_FLAG_TOGGLE_TO_1,
-                             &hits, &misses, &excludes);
-         cover_item_to_chain(ctx, item, COV_FLAG_TOGGLE_TO_0,
-                             &hits, &misses, &excludes);
-         *skipped += cover_append_to_chain(&(ctx->ch_toggle), item, line,
-                                           hits, misses, excludes, limit);
-         break;
-
-      case COV_ITEM_EXPRESSION:
-         if (item->flags & COV_FLAG_00)
-            cover_item_to_chain(ctx, item, COV_FLAG_00,
-                                &hits, &misses, &excludes);
-         if (item->flags & COV_FLAG_01)
-            cover_item_to_chain(ctx, item, COV_FLAG_01,
-                                &hits, &misses, &excludes);
-         if (item->flags & COV_FLAG_10)
-            cover_item_to_chain(ctx, item, COV_FLAG_10,
-                                &hits, &misses, &excludes);
-         if (item->flags & COV_FLAG_11)
-            cover_item_to_chain(ctx, item, COV_FLAG_11,
-                                &hits, &misses, &excludes);
-         if (item->flags & COV_FLAG_TRUE)
-            cover_item_to_chain(ctx, item, COV_FLAG_TRUE,
-                                &hits, &misses, &excludes);
-         if (item->flags & COV_FLAG_FALSE)
-            cover_item_to_chain(ctx, item, COV_FLAG_FALSE,
-                                &hits, &misses, &excludes);
-
-         *skipped += cover_append_to_chain(&(ctx->ch_expression), item, line,
-                                           hits, misses, excludes, limit);
-         break;
-
-      case COV_ITEM_STATE:
-         cover_item_to_chain(ctx, item, COV_FLAG_STATE, &hits, &misses, &excludes);
-         *skipped += cover_append_to_chain(&(ctx->ch_state), item, line,
-                                           hits, misses, excludes, limit);
-         break;
-
-      default:
-         fatal("unsupported type of code coverage: %d !", item->kind);
-      }
-   }
-
-   cover_report_children(ctx, s, dir, summf, skipped);
-}
-
-static void cover_report_hierarchy(cover_report_ctx_t *ctx,
-                                   cover_scope_t *s, const char *dir)
-{
-   char *hier LOCAL = xasprintf("%s/%s.html", dir, istr(s->hier));
-
-   // TODO: Handle escaped identifiers in hierarchy path!
-   FILE *f = fopen(hier, "w");
-   if (f == NULL)
-      fatal("failed to open report file: %s\n", hier);
-
-   INIT_CHAIN(ctx, ch_stmt);
-   INIT_CHAIN(ctx, ch_branch);
-   INIT_CHAIN(ctx, ch_toggle);
-   INIT_CHAIN(ctx, ch_expression);
-   INIT_CHAIN(ctx, ch_state);
-
-   cover_print_html_header(f, ctx, false, s, "NVC code coverage report");
-
-   fprintf(f, "<h2 style=\"margin-left: " MARGIN_LEFT ";\">\n  Sub-instances:\n</h2>\n\n");
-   cover_print_hierarchy_header(f);
-
-   int skipped = 0;
-   cover_report_children(ctx, s, dir, f, &skipped);
-
-   cover_print_hierarchy_footer(f);
-
-   fprintf(f, "<h2 style=\"margin-left: " MARGIN_LEFT ";\">\n  Current Instance:\n</h2>\n\n");
-   cover_print_hierarchy_header(f);
-   cover_print_hierarchy_summary(f, ctx, s->hier, false, true, true);
-   cover_print_hierarchy_footer(f);
-
-   fprintf(f, "<h2 style=\"margin-left: " MARGIN_LEFT ";\">\n  Details:\n</h2>\n\n");
-   if (skipped)
-      fprintf(f, "<h3 style=\"margin-left: " MARGIN_LEFT ";\">The limit of "
-                 "printed items was reached (%d). Total %d items are not "
-                 "displayed.</h3>\n\n", ctx->data->report_item_limit, skipped);
-   cover_print_hierarchy_guts(f, ctx);
-   cover_print_timestamp(f);
-
-   fclose(f);
-}
-
-static void cover_report_children(cover_report_ctx_t *ctx,
-                                  cover_scope_t *s, const char *dir,
-                                  FILE *summf, int *skipped)
-{
-   for (list_iter(cover_scope_t *, it, s->children)) {
-      if (it->type == CSCOPE_INSTANCE) {
-         // Collect coverage of sub-block
-         cover_report_ctx_t sub_ctx = {};
-         sub_ctx.parent = ctx;
-         sub_ctx.data = ctx->data;
-         sub_ctx.lvl = ctx->lvl + 2;
-
-         cover_report_hierarchy(&sub_ctx, it, dir);
-
-         cover_print_hierarchy_summary(summf, &sub_ctx,
-                                       it->hier, false, false, false);
-
-         // Add coverage from sub-hierarchies
-         ctx->nested_stats.hit_stmts += sub_ctx.nested_stats.hit_stmts;
-         ctx->nested_stats.total_stmts += sub_ctx.nested_stats.total_stmts;
-         ctx->nested_stats.hit_branches += sub_ctx.nested_stats.hit_branches;
-         ctx->nested_stats.total_branches += sub_ctx.nested_stats.total_branches;
-         ctx->nested_stats.hit_toggles += sub_ctx.nested_stats.hit_toggles;
-         ctx->nested_stats.total_toggles += sub_ctx.nested_stats.total_toggles;
-         ctx->nested_stats.hit_expressions += sub_ctx.nested_stats.hit_expressions;
-         ctx->nested_stats.total_expressions += sub_ctx.nested_stats.total_expressions;
-         ctx->nested_stats.hit_states += sub_ctx.nested_stats.hit_states;
-         ctx->nested_stats.total_states += sub_ctx.nested_stats.total_states;
-      }
-      else
-         cover_report_scope(ctx, it, dir, summf, skipped);
-   }
-}
-
-void cover_report(const char *path, cover_data_t *data, int item_limit)
-{
-   char *subdir LOCAL = xasprintf("%s/hier", path);
-   make_dir(path);
-   make_dir(subdir);
-
-   static const struct {
-      const char *name;
-      cover_mask_t mask;
-   } lst[] = {
-      { "covered",      COVER_MASK_DONT_PRINT_COVERED       },
-      { "uncovered",    COVER_MASK_DONT_PRINT_UNCOVERED     },
-      { "excluded",     COVER_MASK_DONT_PRINT_EXCLUDED      },
-   };
-
-   LOCAL_TEXT_BUF tb = tb_new();
-   tb_printf(tb, "Code coverage report contains: ");
-
-   bool first = true;
-   for (int i = 0; i < ARRAY_LEN(lst); i++)
-      if (!cover_enabled(data, lst[i].mask)) {
-         if (first)
-            first = false;
-         else
-            tb_printf(tb, ", ");
-         tb_printf(tb, "%s", lst[i].name);
-      }
-   tb_printf(tb, " coverage details.");
-
-   notef("Code coverage report folder: %s.", path);
-   notef("%s", tb_get(tb));
-
-   cover_scope_t *child0 = list_get(data->root_scope->children, 0);
-
-   cover_report_ctx_t top_ctx = {};
-   top_ctx.data = data;
-   data->report_item_limit = item_limit;
-   cover_report_hierarchy(&top_ctx, child0, subdir);
-
-   char *top LOCAL = xasprintf("%s/index.html", path);
-   FILE *f = fopen(top, "w");
-
-   cover_print_html_header(f, &top_ctx, true, child0, "NVC code coverage report");
-   cover_print_hierarchy_header(f);
-   cover_print_hierarchy_summary(f, &top_ctx, child0->hier, true, true, false);
-   cover_print_hierarchy_footer(f);
-   cover_print_timestamp(f);
-
-   if (opt_get_int(OPT_VERBOSE)) {
-      notef("Coverage for sub-hierarchies:");
-      printf("%-55s %-20s %-20s %-20s %-20s %-20s\n",
-             "Hierarchy", "Statement", "Branch", "Toggle", "Expression", "FSM state");
-      cover_rpt_buf_t *buf = data->rpt_buf;
-      while (buf) {
-         printf("%s\n", tb_get(buf->tb));
-         tb_free(buf->tb);
-         data->rpt_buf = buf->prev;
-         free(buf);
-         buf = data->rpt_buf;
-      };
-   }
-
-   fclose(f);
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Cobertura XML export
-
-typedef struct _cobertura_class cobertura_class_t;
-
-typedef struct {
-   unsigned lineno;
-   unsigned hits;
-   bool     branch;
-   unsigned bflags;
-} cobertura_line_t;
-
-typedef struct _cobertura_class {
-   const char        *file;
-   ident_t            name;
-   cobertura_class_t *next;
-   unsigned           nlines;
-   unsigned           maxlines;
-   cobertura_line_t  *lines;
-} cobertura_class_t;
-
-typedef struct {
-   hash_t            *class_map;
-   cobertura_class_t *classes;
-   char              *relative;
-} cobertura_report_t;
-
-static cobertura_class_t *cobertura_get_class(cobertura_report_t *report,
-                                              ident_t name, const loc_t *loc)
-{
-   cobertura_class_t *c = hash_get(report->class_map, name);
-   if (c == NULL) {
-      c = xcalloc(sizeof(cobertura_class_t));
-      c->name = name;
-      c->file = loc_file_str(loc);
-      c->next = report->classes;
-
-      if (is_absolute_path(c->file) && report->relative != NULL) {
-         const size_t rlen = strlen(report->relative);
-         if (strncmp(report->relative, c->file, rlen) == 0) {
-            c->file += rlen;
-            while (c->file[0] == DIR_SEP[0] || c->file[0] == '/')
-               c->file++;
-         }
-      }
-
-      report->classes = c;
-      hash_put(report->class_map, name, c);
-   }
-
-   return c;
-}
-
-static cobertura_line_t *cobertura_get_line(cobertura_class_t *class,
-                                            const loc_t *loc)
-{
-   if (class->nlines > 0) {
-      cobertura_line_t *last = &(class->lines[class->nlines - 1]);
-      if (last->lineno == loc->first_line)  // Most likely
-         return last;
-
-      for (int i = 0; i < class->nlines - 1; i++) {
-         cobertura_line_t *line = &(class->lines[i]);
-         if (line->lineno == loc->first_line)
-            return line;
-      }
-   }
-
-   if (class->nlines == class->maxlines) {
-      class->maxlines = MAX(class->maxlines * 2, 100);
-      class->lines = xrealloc_array(class->lines, class->maxlines,
-                                    sizeof(cobertura_line_t));
-   }
-
-   cobertura_line_t *line = &(class->lines[class->nlines++]);
-   memset(line, '\0', sizeof(cobertura_line_t));
-   line->lineno = loc->first_line;
-   return line;
-}
-
-static void cobertura_export_scope(cobertura_report_t *report,
-                                   cobertura_class_t *class,
-                                   cover_scope_t *s)
-{
-   if (s->block_name != NULL)
-      class = cobertura_get_class(report, s->block_name, &s->loc);
-
-   for (int i = 0; i < s->items.count; i++) {
-      cover_item_t *t = &(s->items.items[i]);
-      switch (t->kind) {
-      case COV_ITEM_STMT:
-         {
-            cobertura_line_t *l = cobertura_get_line(class, &(t->loc));
-            l->hits += t->data;
-         }
-         break;
-      case COV_ITEM_BRANCH:
-         {
-            cobertura_line_t *l = cobertura_get_line(class, &(t->loc));
-            l->branch = true;
-            l->bflags |= t->data;
-         }
-         break;
-      default:
-         break;
-      }
-   }
-
-   for (list_iter(cover_scope_t *, it, s->children))
-      cobertura_export_scope(report, class, it);
-}
-
-static void cobertura_class_stats(const cobertura_class_t *class,
-                                  int *nlines, int *hitlines,
-                                  int *nbranches, int *hitbranches)
-{
-   *nlines += class->nlines;
-   for (int i = 0; i < class->nlines; i++) {
-      const cobertura_line_t *line = &(class->lines[i]);
-      if (line->hits > 0)
-         (*hitlines)++;
-      if (line->branch) {
-         (*nbranches)++;
-         if ((line->bflags & COV_FLAG_TRUE) && (line->bflags & COV_FLAG_FALSE))
-             (*hitbranches)++;
-      }
-   }
-}
-
-static void cobertura_print_class(cobertura_class_t *class, FILE *f)
-{
-   ident_t ename = ident_until(class->name, '-');
-   ident_t aname = ident_from(class->name, '-');
-
-   int nlines = 0, hitlines = 0, nbranches = 0, hitbranches = 0;
-   cobertura_class_stats(class, &nlines, &hitlines, &nbranches, &hitbranches);
-
-   const double line_rate = (double)hitlines / (double)nlines;
-   const double branch_rate = (double)hitbranches / (double)nbranches;
-
-   fprintf(f, "<class name=\"%s(%s)\" filename=\"%s\" "
-           "line-rate=\"%f\" branch-rate=\"%f\" complexity=\"0.0\" >\n",
-           istr(ename), istr(aname), class->file, line_rate, branch_rate);
-   fprintf(f, "<methods/>\n");
-
-   fprintf(f, "<lines>\n");
-   for (int i = 0; i < class->nlines; i++) {
-      const cobertura_line_t *line = &(class->lines[i]);
-      if (line->branch) {
-         int pct = 0;
-         if (line->bflags & COV_FLAG_TRUE) pct += 50;
-         if (line->bflags & COV_FLAG_FALSE) pct += 50;
-
-         fprintf(f, "<line number=\"%d\" hits=\"%d\" branch=\"true\" "
-                 "condition-coverage=\"%d %%\">\n",
-                 line->lineno, line->hits, pct);
-         fprintf(f, "<conditions>\n");
-         fprintf(f, "<condition number=\"0\" type=\"jump\" "
-                 "coverage=\"%d %%\"/>\n", pct);
-         fprintf(f, "</conditions>\n");
-         fprintf(f, "</line>\n");
-      }
-      else
-         fprintf(f, "<line number=\"%d\" hits=\"%d\" branch=\"false\"/>\n",
-                 line->lineno, line->hits);
-   }
-   fprintf(f, "</lines>\n");
-
-   fprintf(f, "</class>\n");
-}
-
-void cover_export_cobertura(cover_data_t *data, FILE *f, const char *relative)
-{
-   cobertura_report_t report = {
-      .class_map = hash_new(64),
-      .relative = relative ? realpath(relative, NULL) : NULL,
-   };
-
-   cobertura_export_scope(&report, NULL, data->root_scope);
-
-   fprintf(f, "<?xml version='1.0' encoding='UTF-8'?>\n");
-   fprintf(f, "<!DOCTYPE coverage SYSTEM "
-           "'http://cobertura.sourceforge.net/xml/coverage-04.dtd'>\n");
-
-   int nlines = 0, hitlines = 0, nbranches = 0, hitbranches = 0;
-   for (cobertura_class_t *it = report.classes; it; it = it->next)
-      cobertura_class_stats(it, &nlines, &hitlines, &nbranches, &hitbranches);
-
-   const double line_rate = (double)hitlines / (double)nlines;
-   const double branch_rate = (double)hitbranches / (double)nbranches;
-
-   unsigned long timestamp;
-   const long override_time = opt_get_int(OPT_COVER_TIMESTAMP);
-   if (override_time >= 0)
-      timestamp = override_time;
-   else
-      timestamp = time(NULL);
-
-   fprintf(f, "<coverage version=\"" PACKAGE_STRING "\" "
-           "line-rate=\"%f\" branch-rate=\"%f\" complexity=\"0.0\" "
-           "lines-valid=\"%d\" lines-covered=\"%d\" "
-           "branches-valid=\"%d\" branches-covered=\"%d\" "
-           "timestamp=\"%lu\">\n",
-           line_rate, branch_rate, nlines, hitlines, nbranches, hitbranches,
-           timestamp);
-   fprintf(f, "<sources>\n");
-   fprintf(f, "<source>.</source>\n");
-   fprintf(f, "</sources>\n");
-   fprintf(f, "<packages>\n");
-   fprintf(f, "<package name=\"%s\" "
-           "line-rate=\"%f\" branch-rate=\"%f\" complexity=\"0.0\">\n",
-           istr(data->root_scope->name), line_rate, branch_rate);
-
-   fprintf(f, "<classes>\n");
-   for (cobertura_class_t *it = report.classes; it; it = it->next)
-      cobertura_print_class(it, f);
-   fprintf(f, "</classes>\n");
-
-   fprintf(f, "</package>\n");
-   fprintf(f, "</packages>\n");
-   fprintf(f, "</coverage>\n");
-
-   for (cobertura_class_t *it = report.classes, *tmp; it; it = tmp) {
-      tmp = it->next;
-      free(it->lines);
-      free(it);
-   }
-
-   free(report.relative);
-   hash_free(report.class_map);
 }
